Refactor streams controller for better audio bufering
