+++ music_assistant/controllers/streams/__init__.py
@@ -7,6 +7,6 @@ streams for playback by players as well as fades between songs.
 
 from __future__ import annotations
 
-from .streams_controller import StreamsController
+from .controller import StreamsController
 
 __all__ = ["StreamsController"]
