+++ music_assistant/controllers/streams/audio.py
@@ -1,0 +1,2108 @@
+"""
+Audio streaming helpers that interact with core controllers and providers.
+
+This module contains all audio stream acquisition and processing functions
+that need access to the MusicAssistant instance. Generic audio utilities
+that do not need controller interaction live in helpers/audio.py.
+"""
+
+from __future__ import annotations
+
+import asyncio
+import logging
+import os
+import re
+import time
+from collections.abc import AsyncGenerator
+from contextlib import asynccontextmanager
+from dataclasses import dataclass
+from functools import partial
+from typing import TYPE_CHECKING, Any, cast
+
+import aiofiles
+import aiohttp
+import shortuuid
+from aiohttp import ClientConnectorSSLError, ClientTimeout
+from music_assistant_models.dsp import DSPConfig, DSPDetails, DSPState
+from music_assistant_models.enums import (
+    ContentType,
+    MediaType,
+    PlayerFeature,
+    PlayerType,
+    StreamType,
+    VolumeNormalizationMode,
+)
+from music_assistant_models.errors import (
+    AudioError,
+    InvalidDataError,
+    MediaNotFoundError,
+    MusicAssistantError,
+    ProviderPermissionDenied,
+    ProviderUnavailableError,
+    QueueEmpty,
+    RetriesExhausted,
+)
+from music_assistant_models.media_items import AudioFormat, Track
+from music_assistant_models.player_queue import PlayLogEntry
+from music_assistant_models.streamdetails import StreamMetadata
+
+from music_assistant.constants import (
+    CONF_CROSSFADE_DURATION,
+    CONF_ENTRY_OUTPUT_LIMITER,
+    CONF_ENTRY_SUPPORT_GAPLESS_DIFFERENT_SAMPLE_RATES,
+    CONF_ENTRY_VOLUME_NORMALIZATION_TARGET,
+    CONF_OUTPUT_CHANNELS,
+    CONF_SAMPLE_RATES,
+    CONF_SMART_FADES_MODE,
+    CONF_VOLUME_NORMALIZATION_FIXED_GAIN_RADIO,
+    CONF_VOLUME_NORMALIZATION_FIXED_GAIN_TRACKS,
+    CONF_VOLUME_NORMALIZATION_TARGET,
+    INTERNAL_PCM_FORMAT,
+    MASS_LOGGER_NAME,
+    VERBOSE_LOG_LEVEL,
+)
+from music_assistant.controllers.streams.audio_buffer import AudioBuffer
+from music_assistant.controllers.streams.constants import (
+    CACHE_CATEGORY_RESOLVED_RADIO_URL,
+    CACHE_PROVIDER,
+    CONF_ALLOW_CROSSFADE_SAME_ALBUM,
+    CONF_BUFFER_SIZE,
+    CONF_BUFFER_SIZE_DEFAULT,
+    BufferSize,
+)
+from music_assistant.controllers.streams.ogg_handler import get_chained_ogg_stream
+from music_assistant.controllers.streams.smart_fades import SmartFadesMixer
+from music_assistant.controllers.streams.smart_fades.fades import SMART_CROSSFADE_DURATION
+from music_assistant.helpers import ssl as ssl_util
+from music_assistant.helpers.audio import (
+    HTTP_HEADERS,
+    HTTP_HEADERS_ICY,
+    _get_normalization_mode,
+    _get_parts_from_position,
+    get_bit_rate,
+    get_chunksize,
+    is_grouping_preventing_dsp,
+    parse_extinf_metadata,
+    resample_pcm_audio,
+)
+from music_assistant.helpers.dsp import filter_to_ffmpeg_params
+from music_assistant.helpers.ffmpeg import FFMpeg, get_ffmpeg_stream
+from music_assistant.helpers.playlists import IsHLSPlaylist, PlaylistItem, fetch_playlist, parse_m3u
+from music_assistant.helpers.throttle_retry import BYPASS_THROTTLER
+from music_assistant.helpers.util import (
+    clean_stream_title,
+    detect_charset,
+    divide_chunks,
+    remove_file,
+)
+from music_assistant.models.smart_fades import SmartFadesMode
+from music_assistant.providers.sync_group.constants import SGP_PREFIX
+
+if TYPE_CHECKING:
+    from music_assistant_models.player_queue import PlayerQueue
+    from music_assistant_models.queue_item import QueueItem
+    from music_assistant_models.streamdetails import StreamDetails
+
+    from music_assistant.mass import MusicAssistant
+    from music_assistant.models.music_provider import MusicProvider
+    from music_assistant.models.player import Player
+    from music_assistant.providers.sync_group import SyncGroupPlayer
+
+# ruff: noqa: PLR0915
+
+
+@dataclass
+class CrossfadeData:
+    """Data class to hold crossfade data."""
+
+    data: bytes
+    fade_in_size: int
+    pcm_format: AudioFormat  # Format of the 'data' bytes (current/previous track's format)
+    fade_in_pcm_format: AudioFormat  # Format for 'fade_in_size' (next track's format)
+    queue_item_id: str
+
+
+class StreamsAudio:
+    """Audio stream acquisition and processing for the streams controller."""
+
+    def __init__(self, mass: MusicAssistant) -> None:
+        """
+        Initialize StreamsAudio.
+
+        :param mass: The MusicAssistant instance.
+        """
+        self.mass = mass
+        self.logger = logging.getLogger(f"{MASS_LOGGER_NAME}.streams.audio")
+        self._crossfade_data: dict[str, CrossfadeData] = {}
+        self._smart_fades_mixer = SmartFadesMixer(mass.streams)
+
+    # --- Public methods ---
+
+    async def get_stream_details(
+        self,
+        queue_item: QueueItem,
+        seek_position: int = 0,
+        fade_in: bool = False,
+        prefer_album_loudness: bool = False,
+    ) -> StreamDetails:
+        """
+        Get streamdetails for the given QueueItem.
+
+        This is called just-in-time when a PlayerQueue wants a MediaItem to be played.
+        Do not try to request streamdetails too much in advance as this is expiring data.
+        """
+        mass = self.mass
+        streamdetails: StreamDetails | None = None
+        time_start = time.time()
+        self.logger.debug("Getting streamdetails for %s", queue_item.uri)
+
+        if not queue_item.media_item and not queue_item.streamdetails:
+            # in case of a non-media item queue item, the streamdetails should already be provided
+            # this should not happen, but guard it just in case
+            raise MediaNotFoundError(
+                f"Unable to retrieve streamdetails for {queue_item.name} ({queue_item.uri})"
+            )
+
+        existing_buffer = None
+        if queue_item.streamdetails and (
+            (queue_item.streamdetails.created_at + queue_item.streamdetails.expiration)
+            > time.time()
+            or (
+                (existing_buffer := queue_item.streamdetails.buffer)
+                and existing_buffer.is_valid(seek_position * 1000)
+            )
+        ):
+            # already got a fresh/unused (or unexpired) streamdetails
+            streamdetails = queue_item.streamdetails
+        else:
+            # need to (re)create streamdetails
+            # retrieve streamdetails from provider
+
+            media_item = queue_item.media_item
+            assert media_item is not None  # for type checking
+            preferred_providers: list[str] = []
+            if (
+                (queue := mass.player_queues.get(queue_item.queue_id))
+                and queue.userid
+                and (playback_user := await mass.webserver.auth.get_user(queue.userid))
+                and playback_user.provider_filter
+            ):
+                # handle steering into user preferred providerinstance
+                preferred_providers = playback_user.provider_filter
+            else:
+                preferred_providers = [x.provider_instance for x in media_item.provider_mappings]
+            for allow_other_provider in (False, True):
+                if streamdetails:
+                    break
+                # sort by quality and check item's availability
+                for prov_media in sorted(
+                    media_item.provider_mappings, key=lambda x: x.quality or 0, reverse=True
+                ):
+                    if not prov_media.available:
+                        self.logger.debug(f"Skipping unavailable {prov_media}")
+                        continue
+                    if (
+                        not allow_other_provider
+                        and prov_media.provider_instance not in preferred_providers
+                    ):
+                        continue
+                    # guard that provider is available
+                    music_prov = mass.get_provider(prov_media.provider_instance)
+                    if TYPE_CHECKING:  # avoid circular import
+                        assert isinstance(music_prov, MusicProvider)
+                    if not music_prov:
+                        self.logger.debug(f"Skipping {prov_media} - provider not available")
+                        continue  # provider not available ?
+                    # get streamdetails from provider
+                    try:
+                        BYPASS_THROTTLER.set(True)
+                        streamdetails = await music_prov.get_stream_details(
+                            prov_media.item_id, media_item.media_type
+                        )
+                    except MusicAssistantError as err:
+                        self.logger.warning(str(err))
+                    else:
+                        break
+                    finally:
+                        BYPASS_THROTTLER.set(False)
+
+            if not streamdetails:
+                msg = f"Unable to retrieve streamdetails for {queue_item.name} ({queue_item.uri})"
+                raise MediaNotFoundError(msg)
+
+            # work out how to handle radio stream
+            if (
+                streamdetails.stream_type in (StreamType.ICY, StreamType.HLS, StreamType.HTTP)
+                and streamdetails.media_type == MediaType.RADIO
+                and isinstance(streamdetails.path, str)
+            ):
+                resolved_url, stream_type = await self.resolve_radio_stream(streamdetails.path)
+                streamdetails.path = resolved_url
+                streamdetails.stream_type = stream_type
+                # Set up metadata monitoring callback for HLS radio streams, if not already set
+                if (
+                    stream_type == StreamType.HLS
+                    and not streamdetails.stream_metadata_update_callback
+                ):
+                    streamdetails.stream_metadata_update_callback = partial(
+                        self._update_hls_radio_metadata
+                    )
+                    streamdetails.stream_metadata_update_interval = 5
+            # handle volume normalization details
+            if result := await mass.music.get_loudness(
+                streamdetails.item_id, streamdetails.provider, media_type=queue_item.media_type
+            ):
+                streamdetails.loudness = result[0]
+                streamdetails.loudness_album = result[1]
+
+        if not streamdetails.duration:
+            streamdetails.duration = queue_item.duration
+        if seek_position and (not streamdetails.allow_seek or not streamdetails.duration):
+            self.logger.warning("seeking is not possible on duration-less streams!")
+            seek_position = 0
+
+        # set queue_id on the streamdetails so we know what is being streamed
+        streamdetails.queue_id = queue_item.queue_id
+        # handle skip/fade_in details
+        streamdetails.seek_position = seek_position
+        streamdetails.fade_in = fade_in
+
+        streamdetails.prefer_album_loudness = prefer_album_loudness
+        player_settings = await mass.config.get_player_config(streamdetails.queue_id)
+        core_config = await mass.config.get_core_config("streams")
+        conf_volume_normalization_target = float(
+            str(player_settings.get_value(CONF_VOLUME_NORMALIZATION_TARGET, -17))
+        )
+        # guard against invalid volume normalization values
+        # range and default_value are guaranteed to be set for this constant
+        volume_range = CONF_ENTRY_VOLUME_NORMALIZATION_TARGET.range
+        assert volume_range is not None
+        if (
+            conf_volume_normalization_target < volume_range[0]
+            or conf_volume_normalization_target >= volume_range[1]
+        ):
+            default_val = CONF_ENTRY_VOLUME_NORMALIZATION_TARGET.default_value
+            assert isinstance(default_val, (int, float))
+            conf_volume_normalization_target = float(default_val)
+            self.logger.warning(
+                "Invalid volume normalization target configured for player %s, "
+                "resetting to default of %s dB",
+                streamdetails.queue_id,
+                CONF_ENTRY_VOLUME_NORMALIZATION_TARGET.default_value,
+            )
+        streamdetails.target_loudness = conf_volume_normalization_target
+        streamdetails.volume_normalization_mode = _get_normalization_mode(
+            core_config, player_settings, streamdetails
+        )
+
+        # attach the DSP details of all group members
+        streamdetails.dsp = self.get_stream_dsp_details(streamdetails.queue_id)
+
+        self.logger.debug(
+            "retrieved streamdetails for %s in %s milliseconds",
+            queue_item.uri,
+            int((time.time() - time_start) * 1000),
+        )
+        return streamdetails
+
+    async def get_media_stream(
+        self,
+        streamdetails: StreamDetails,
+        pcm_format: AudioFormat,
+        seek_position: int = 0,
+        filter_params: list[str] | None = None,
+    ) -> AsyncGenerator[bytes, None]:
+        """Get audio stream for given media details as raw PCM."""
+        mass = self.mass
+        logger = self.logger.getChild("media_stream")
+        logger.log(VERBOSE_LOG_LEVEL, "Starting media stream for %s", streamdetails.uri)
+        extra_input_args = streamdetails.extra_input_args or []
+
+        # work out audio source for these streamdetails
+        audio_source: str | AsyncGenerator[bytes, None]
+        stream_type = streamdetails.stream_type
+        if stream_type == StreamType.CUSTOM:
+            music_prov = mass.get_provider(streamdetails.provider)
+            if TYPE_CHECKING:  # avoid circular import
+                assert isinstance(music_prov, MusicProvider)
+            audio_source = music_prov.get_audio_stream(
+                streamdetails, seek_position=seek_position if streamdetails.can_seek else 0
+            )
+            seek_position = 0 if streamdetails.can_seek else seek_position
+        elif stream_type == StreamType.ICY:
+            assert isinstance(streamdetails.path, str)  # for type checking
+            audio_source = self.get_icy_radio_stream(streamdetails.path, streamdetails)
+            seek_position = 0  # seeking not possible on radio streams
+        elif stream_type == StreamType.IN_BAND:
+            assert isinstance(streamdetails.path, str)  # for type checking
+
+            # For IN_BAND (OGG/Opus) radio streams, use chained OGG handler.
+            # This handles the chained OGG format by stitching logical bitstreams together
+            # so FFmpeg sees a single continuous stream. Metadata is extracted in-band.
+            def _on_inband_metadata(metadata: dict[str, str]) -> None:
+                """Handle metadata extracted from the OGG stream."""
+                title = metadata.get("title", "")
+                artist = metadata.get("artist", "")
+                album = metadata.get("album", "")
+                if not artist and " - " in title:
+                    artist, title = title.split(" - ", 1)
+                if title or artist:
+                    if artist and title:
+                        stream_title = f"{artist} - {title}"
+                    elif title:
+                        stream_title = title
+                    else:
+                        stream_title = artist
+                    cleaned_title = clean_stream_title(stream_title)
+                    if cleaned_title and cleaned_title != streamdetails.stream_title:
+                        self.logger.log(VERBOSE_LOG_LEVEL, "In-band metadata: %s", cleaned_title)
+                        streamdetails.stream_title = cleaned_title
+                        streamdetails.stream_metadata = StreamMetadata(
+                            title=title or cleaned_title, artist=artist or None, album=album or None
+                        )
+
+            audio_source = get_chained_ogg_stream(
+                mass, streamdetails.path, metadata_callback=_on_inband_metadata
+            )
+            seek_position = 0  # seeking not possible on radio streams
+        elif stream_type == StreamType.HLS:
+            assert isinstance(streamdetails.path, str)  # for type checking
+            substream = await self.get_hls_substream(streamdetails.path)
+            audio_source = substream.path
+            if streamdetails.media_type == MediaType.RADIO:
+                # HLS streams (especially the BBC) struggle when they're played directly
+                # with ffmpeg, where they just stop after some minutes,
+                # so we tell ffmpeg to loop around in this case.
+                extra_input_args += ["-stream_loop", "-1", "-re"]
+        else:
+            # all other stream types (HTTP, FILE, etc)
+            if stream_type == StreamType.ENCRYPTED_HTTP:
+                assert streamdetails.decryption_key is not None  # for type checking
+                extra_input_args += ["-decryption_key", streamdetails.decryption_key]
+            if isinstance(streamdetails.path, list):
+                # multi part stream
+                audio_source = self.get_multi_file_stream(streamdetails, seek_position)
+                seek_position = 0  # handled by get_multi_file_stream
+            else:
+                # regular single file/url stream
+                assert isinstance(streamdetails.path, str)  # for type checking
+                audio_source = streamdetails.path
+
+        # handle seek support
+        if seek_position and streamdetails.duration and streamdetails.allow_seek:
+            extra_input_args += ["-ss", str(int(seek_position))]
+
+        bytes_sent = 0
+        finished = False
+        cancelled = False
+        first_chunk_received = False
+        ffmpeg_loglevel = "debug" if self.logger.isEnabledFor(VERBOSE_LOG_LEVEL) else "info"
+        ffmpeg_proc = FFMpeg(
+            audio_input=audio_source,
+            input_format=streamdetails.audio_format,
+            output_format=pcm_format,
+            filter_params=filter_params,
+            extra_input_args=extra_input_args,
+            collect_log_history=True,
+            loglevel=ffmpeg_loglevel,
+        )
+
+        try:
+            await ffmpeg_proc.start()
+            assert ffmpeg_proc.proc is not None  # for type checking
+            logger.debug(
+                "Started media stream for %s - using streamtype: %s "
+                "- pcm format: %s - ffmpeg PID: %s",
+                streamdetails.uri,
+                streamdetails.stream_type,
+                pcm_format.content_type.value,
+                ffmpeg_proc.proc.pid,
+            )
+            stream_start = mass.loop.time()
+            chunk_size = get_chunksize(pcm_format, 1)
+            async for chunk in ffmpeg_proc.iter_chunked(chunk_size):
+                if not first_chunk_received:
+                    # At this point ffmpeg has started and should now know the codec used
+                    # for encoding the audio.
+                    first_chunk_received = True
+                    streamdetails.audio_format.codec_type = ffmpeg_proc.input_format.codec_type
+                    logger.debug(
+                        "First chunk received after %.2f seconds (codec detected: %s)",
+                        mass.loop.time() - stream_start,
+                        ffmpeg_proc.input_format.codec_type,
+                    )
+                yield chunk
+                bytes_sent += len(chunk)
+
+            # end of audio/track reached
+            logger.debug("End of ffmpeg output stream reached for %s", streamdetails.uri)
+            # wait until stderr also completed reading
+            await ffmpeg_proc.wait_with_timeout(5)
+            logger.debug(
+                "FFmpeg process ended with return code %s for %s",
+                ffmpeg_proc.returncode,
+                streamdetails.uri,
+            )
+            if ffmpeg_proc.returncode not in (0, None):
+                log_trail = "\n".join(list(ffmpeg_proc.log_history)[-5:])
+                raise AudioError(f"FFMpeg exited with code {ffmpeg_proc.returncode}: {log_trail}")
+            if bytes_sent == 0:
+                # edge case: no audio data was received at all
+                raise AudioError("No audio was received")
+            finished = True
+        except (Exception, GeneratorExit, asyncio.CancelledError) as err:
+            if isinstance(err, asyncio.CancelledError | GeneratorExit):
+                # we were cancelled, just raise
+                cancelled = True
+                raise
+            # dump the last 10 lines of the log in case of an unclean exit
+            logger.warning("\n".join(list(ffmpeg_proc.log_history)[-10:]))
+            raise AudioError(f"Error while streaming: {err}") from err
+        finally:
+            # always ensure close is called which also handles all cleanup
+            await ffmpeg_proc.close()
+            # determine how many seconds we've received
+            # for pcm output we can calculate this easily
+            seconds_received = bytes_sent / pcm_format.pcm_sample_size if bytes_sent else 0
+            # store accurate duration
+            if finished and not seek_position and seconds_received:
+                streamdetails.duration = int(seconds_received)
+
+            logger.log(
+                VERBOSE_LOG_LEVEL,
+                "stream %s (with code %s) for %s",
+                "cancelled" if cancelled else "finished" if finished else "aborted",
+                ffmpeg_proc.returncode,
+                streamdetails.uri,
+            )
+
+    async def resolve_radio_stream(self, url: str) -> tuple[str, StreamType]:
+        """
+        Resolve a streaming radio URL.
+
+        Unwraps any playlists if needed.
+        Determines if the stream supports ICY metadata or in-band metadata.
+
+        Returns tuple;
+        - unfolded URL as string
+        - StreamType to determine ICY (radio), HLS, or IN_BAND stream.
+        """
+        mass = self.mass
+        if cache := await mass.cache.get(
+            key=url, provider=CACHE_PROVIDER, category=CACHE_CATEGORY_RESOLVED_RADIO_URL
+        ):
+            if TYPE_CHECKING:  # for type checking
+                cache = cast("tuple[str, str]", cache)
+            return (cache[0], StreamType(cache[1]))
+        stream_type = StreamType.HTTP
+        resolved_url = url
+        timeout = ClientTimeout(total=None, connect=10, sock_read=5)
+        try:
+            async with self._connect_radio_stream(
+                url, headers=HTTP_HEADERS_ICY, allow_redirects=True, timeout=timeout
+            ) as resp:
+                headers = resp.headers
+                resp.raise_for_status()
+                if not resp.headers:
+                    raise InvalidDataError("no headers found")
+            content_type = headers.get("content-type", "")
+            if headers.get("icy-metaint") is not None:
+                stream_type = StreamType.ICY
+            elif content_type in ("application/ogg", "audio/ogg"):
+                # Ogg streams (Opus/Vorbis) have in-band metadata via Vorbis comments
+                stream_type = StreamType.IN_BAND
+            if (
+                url.endswith((".m3u", ".m3u8", ".pls"))
+                or ".m3u?" in url
+                or ".m3u8?" in url
+                or ".pls?" in url
+                or "audio/x-mpegurl" in headers.get("content-type", "")
+                or "audio/x-scpls" in headers.get("content-type", "")
+            ):
+                # url is playlist, we need to unfold it
+                try:
+                    substreams = await fetch_playlist(mass, url)
+                    if not any(x for x in substreams if x.length):
+                        for line in substreams:
+                            if not line.is_url:
+                                continue
+                            # unfold first url of playlist
+                            return await self.resolve_radio_stream(line.path)
+                        raise InvalidDataError("No content found in playlist")
+                except IsHLSPlaylist:
+                    stream_type = StreamType.HLS
+
+        except Exception as err:
+            self.logger.warning("Error while parsing radio URL %s: %s", url, str(err))
+            return (url, stream_type)
+
+        result = (resolved_url, stream_type)
+        cache_expiration = 3600 * 3
+        await mass.cache.set(
+            url,
+            result,
+            expiration=cache_expiration,
+            provider=CACHE_PROVIDER,
+            category=CACHE_CATEGORY_RESOLVED_RADIO_URL,
+        )
+        return result
+
+    async def get_icy_radio_stream(
+        self, url: str, streamdetails: StreamDetails
+    ) -> AsyncGenerator[bytes, None]:
+        """Get (radio) audio stream from HTTP, including ICY metadata retrieval."""
+        timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
+        self.logger.debug("Start streaming radio with ICY metadata from url %s", url)
+        async with self._connect_radio_stream(
+            url, allow_redirects=True, headers=HTTP_HEADERS_ICY, timeout=timeout
+        ) as resp:
+            headers = resp.headers
+            meta_int = int(headers["icy-metaint"])
+            while True:
+                try:
+                    yield await resp.content.readexactly(meta_int)
+                    meta_byte = await resp.content.readexactly(1)
+                    if meta_byte == b"\x00":
+                        continue
+                    meta_length = ord(meta_byte) * 16
+                    meta_data = await resp.content.readexactly(meta_length)
+                except asyncio.exceptions.IncompleteReadError:
+                    break
+                if not meta_data:
+                    continue
+                meta_data = meta_data.rstrip(b"\0")
+                stream_title_re = re.search(rb"StreamTitle='([^']*)';", meta_data)
+                if not stream_title_re:
+                    continue
+                try:
+                    # in 99% of the cases the stream title is utf-8 encoded
+                    stream_title = stream_title_re.group(1).decode("utf-8")
+                except UnicodeDecodeError:
+                    # fallback to iso-8859-1
+                    stream_title = stream_title_re.group(1).decode("iso-8859-1", errors="replace")
+                cleaned_stream_title = clean_stream_title(stream_title)
+                if cleaned_stream_title != streamdetails.stream_title:
+                    self.logger.log(
+                        VERBOSE_LOG_LEVEL, "ICY Radio streamtitle original: %s", stream_title
+                    )
+                    self.logger.log(
+                        VERBOSE_LOG_LEVEL, "ICY Radio streamtitle cleaned: %s", cleaned_stream_title
+                    )
+                    streamdetails.stream_title = cleaned_stream_title
+
+    async def get_reconnecting_radio_stream(self, url: str) -> AsyncGenerator[bytes, None]:
+        """
+        Yield continuous radio stream data, automatically reconnecting on disconnect.
+
+        :param url: URL of the radio stream.
+        """
+        timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
+        reconnect_count = 0
+        max_reconnects = 1000  # Allow many reconnects for long-running radio
+
+        while reconnect_count <= max_reconnects:
+            try:
+                async with self._connect_radio_stream(
+                    url, allow_redirects=True, headers=HTTP_HEADERS, timeout=timeout
+                ) as resp:
+                    chunk_count = 0
+                    async for chunk in resp.content.iter_any():
+                        chunk_count += 1
+                        yield chunk
+
+                    # Connection closed normally - reconnect
+                    self.logger.debug(
+                        "Radio stream connection closed after %d chunks, reconnecting... "
+                        "(reconnect #%d)",
+                        chunk_count,
+                        reconnect_count,
+                    )
+                    reconnect_count += 1
+                    await asyncio.sleep(0.1)  # Brief delay before reconnect
+
+            except asyncio.CancelledError:
+                self.logger.debug("Radio stream cancelled for %s", url)
+                raise
+            except (
+                aiohttp.ClientConnectionError,
+                aiohttp.ClientPayloadError,
+                aiohttp.ServerDisconnectedError,
+            ) as err:
+                # Transient network errors - retry
+                self.logger.warning("Radio stream error (reconnect #%d): %s", reconnect_count, err)
+                reconnect_count += 1
+                if reconnect_count > max_reconnects:
+                    raise RetriesExhausted(
+                        f"Radio stream failed after {max_reconnects} reconnects: {err}"
+                    ) from err
+                await asyncio.sleep(0.5)
+            except aiohttp.ClientResponseError as err:
+                if err.status == 404:
+                    raise MediaNotFoundError(f"Radio stream not found: {url}") from err
+                if err.status == 403:
+                    raise ProviderPermissionDenied(f"Radio stream access denied: {url}") from err
+                # Other HTTP errors (5xx etc) - could be temporary
+                raise ProviderUnavailableError(
+                    f"Radio stream returned HTTP {err.status}: {err}"
+                ) from err
+
+        self.logger.warning("Radio stream reached max reconnects (%d) for %s", max_reconnects, url)
+
+    async def get_hls_substream(self, url: str) -> PlaylistItem:
+        """Select the (highest quality) HLS substream for given HLS playlist/URL."""
+        mass = self.mass
+        timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
+        # fetch master playlist and select (best) child playlist
+        # https://datatracker.ietf.org/doc/html/draft-pantos-http-live-streaming-19#section-10
+        async with mass.http_session_no_ssl.get(
+            url, allow_redirects=True, headers=HTTP_HEADERS, timeout=timeout
+        ) as resp:
+            resp.raise_for_status()
+            raw_data = await resp.read()
+            encoding = resp.charset or await detect_charset(raw_data)
+            master_m3u_data = raw_data.decode(encoding)
+        substreams = parse_m3u(master_m3u_data)
+        # There is a chance that we did not get a master playlist with subplaylists
+        # but just a single master/sub playlist with the actual audio stream(s)
+        # so we need to detect if the playlist child's contain audio streams or
+        # sub-playlists.
+        if any(
+            x
+            for x in substreams
+            if (x.length or x.path.endswith((".mp4", ".aac")))
+            and not x.path.endswith((".m3u", ".m3u8"))
+        ):
+            return PlaylistItem(path=url, key=substreams[0].key)
+        # sort substreams on best quality (highest bandwidth) when available
+        if any(x for x in substreams if x.stream_info):
+            substreams.sort(
+                key=lambda x: int(
+                    x.stream_info.get("BANDWIDTH", "0") if x.stream_info is not None else 0
+                ),
+                reverse=True,
+            )
+        substream = substreams[0]
+        if not substream.path.startswith("http"):
+            # path is relative, stitch it together
+            base_path = url.rsplit("/", 1)[0]
+            substream.path = base_path + "/" + substream.path
+        return substream
+
+    async def get_http_stream(
+        self,
+        url: str,
+        streamdetails: StreamDetails,
+        seek_position: int = 0,
+        verify_ssl: bool = True,
+    ) -> AsyncGenerator[bytes, None]:
+        """Get audio stream from HTTP."""
+        mass = self.mass
+        self.logger.debug(
+            "Start HTTP stream for %s (seek_position %s)", streamdetails.uri, seek_position
+        )
+        if seek_position:
+            assert streamdetails.duration, "Duration required for seek requests"
+        http_session = mass.http_session if verify_ssl else mass.http_session_no_ssl
+        # try to get filesize with a head request
+        seek_supported = streamdetails.can_seek
+        if seek_position or not streamdetails.size:
+            async with http_session.head(url, allow_redirects=True, headers=HTTP_HEADERS) as resp:
+                resp.raise_for_status()
+                if size := resp.headers.get("Content-Length"):
+                    streamdetails.size = int(size)
+                seek_supported = resp.headers.get("Accept-Ranges") == "bytes"
+        # headers
+        headers = {**HTTP_HEADERS}
+        timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
+        skip_bytes = 0
+        if seek_position and streamdetails.size:
+            assert streamdetails.duration is not None  # for type checking
+            skip_bytes = int(streamdetails.size / streamdetails.duration * seek_position)
+            headers["Range"] = f"bytes={skip_bytes}-{streamdetails.size}"
+
+        # seeking an unknown or container format is not supported due to the (moov) headers
+        if seek_position and (
+            not seek_supported
+            or streamdetails.audio_format.content_type
+            in (ContentType.UNKNOWN, ContentType.M4A, ContentType.M4B)
+        ):
+            self.logger.warning(
+                "Seeking in %s (%s) not possible.",
+                streamdetails.uri,
+                streamdetails.audio_format.output_format_str,
+            )
+            seek_position = 0
+            streamdetails.seek_position = 0
+
+        # start the streaming from http
+        bytes_received = 0
+        async with http_session.get(
+            url, allow_redirects=True, headers=headers, timeout=timeout
+        ) as resp:
+            is_partial = resp.status == 206
+            if seek_position and not is_partial:
+                raise InvalidDataError("HTTP source does not support seeking!")
+            resp.raise_for_status()
+            async for chunk in resp.content.iter_any():
+                bytes_received += len(chunk)
+                yield chunk
+
+        # store size on streamdetails for later use
+        if not streamdetails.size:
+            streamdetails.size = bytes_received
+        self.logger.debug(
+            "Finished HTTP stream for %s (transferred %s/%s bytes)",
+            streamdetails.uri,
+            bytes_received,
+            streamdetails.size,
+        )
+
+    async def get_file_stream(
+        self,
+        filename: str,
+        streamdetails: StreamDetails,
+        seek_position: int = 0,
+    ) -> AsyncGenerator[bytes, None]:
+        """Get audio stream from local accessible file."""
+        if seek_position:
+            assert streamdetails.duration, "Duration required for seek requests"
+        if not streamdetails.size:
+            stat = await asyncio.to_thread(os.stat, filename)
+            streamdetails.size = stat.st_size
+
+        # seeking an unknown or container format is not supported due to the (moov) headers
+        if seek_position and (
+            streamdetails.audio_format.content_type
+            in (ContentType.UNKNOWN, ContentType.M4A, ContentType.M4B, ContentType.MP4)
+        ):
+            self.logger.warning(
+                "Seeking in %s (%s) not possible.",
+                streamdetails.uri,
+                streamdetails.audio_format.output_format_str,
+            )
+            seek_position = 0
+            streamdetails.seek_position = 0
+
+        chunk_size = get_chunksize(streamdetails.audio_format)
+        async with aiofiles.open(streamdetails.data, "rb") as _file:
+            if seek_position:
+                assert streamdetails.duration is not None  # for type checking
+                seek_pos = int((streamdetails.size / streamdetails.duration) * seek_position)
+                await _file.seek(seek_pos)
+            # yield chunks of data from file
+            while True:
+                data = await _file.read(chunk_size)
+                if not data:
+                    break
+                yield data
+
+    async def get_multi_file_stream(
+        self,
+        streamdetails: StreamDetails,
+        seek_position: int = 0,
+    ) -> AsyncGenerator[bytes, None]:
+        """
+        Return audio stream for a concatenation of multiple files.
+
+        Arguments:
+        seek_position: The position to seek to in seconds
+        """
+        if not isinstance(streamdetails.path, list):
+            raise InvalidDataError("Multi-file streamdetails requires a list of MultiPartPath")
+        parts, seek_position = _get_parts_from_position(streamdetails.path, seek_position)
+        files_list = [part.path for part in parts]
+
+        # concat input files
+        temp_file = f"/tmp/{shortuuid.random(20)}.txt"  # noqa: S108
+        async with aiofiles.open(temp_file, "w") as f:
+            for path in files_list:
+                await f.write(f"file '{path}'\n")
+
+        try:
+            async for chunk in get_ffmpeg_stream(
+                audio_input=temp_file,
+                input_format=streamdetails.audio_format,
+                output_format=AudioFormat(
+                    content_type=ContentType.NUT,
+                    sample_rate=streamdetails.audio_format.sample_rate,
+                    bit_depth=streamdetails.audio_format.bit_depth,
+                    channels=streamdetails.audio_format.channels,
+                ),
+                extra_input_args=[
+                    "-safe",
+                    "0",
+                    "-f",
+                    "concat",
+                    "-i",
+                    temp_file,
+                    "-ss",
+                    str(seek_position),
+                ],
+            ):
+                yield chunk
+        finally:
+            await remove_file(temp_file)
+
+    def attach_loudness_analyzer(
+        self,
+        audio_buffer: AudioBuffer,
+        streamdetails: StreamDetails,
+        max_duration_seconds: int = 120,
+    ) -> None:
+        """
+        Attach a loudness measurement job to an AudioBuffer.
+
+        Registers a chunk callback that feeds raw PCM into an FFmpeg ebur128 process.
+        After max_duration_seconds of audio (or EOF), the measurement is finalized
+        and stored via mass.music.set_loudness.
+
+        :param audio_buffer: The AudioBuffer to observe.
+        :param streamdetails: Stream details for the track being buffered.
+        :param max_duration_seconds: Maximum seconds of audio to analyze.
+        """
+        item_id = streamdetails.item_id
+        provider = streamdetails.provider
+        media_type = streamdetails.media_type
+        pcm_format = audio_buffer.pcm_format
+
+        chunk_queue: asyncio.Queue[bytes | None] = asyncio.Queue()
+        chunks_received = 0
+
+        def _on_chunk(position_seconds: int, pcm_data: bytes) -> None:  # noqa: ARG001
+            nonlocal chunks_received
+            if chunks_received >= max_duration_seconds:
+                return
+            if not pcm_data:
+                # EOF
+                chunk_queue.put_nowait(None)
+                return
+            chunks_received += 1
+            chunk_queue.put_nowait(pcm_data)
+            if chunks_received >= max_duration_seconds:
+                # signal we have enough data
+                chunk_queue.put_nowait(None)
+
+        async def _chunk_generator() -> AsyncGenerator[bytes, None]:
+            """Yield chunks from the queue until None (EOF/done)."""
+            while True:
+                chunk = await chunk_queue.get()
+                if chunk is None:
+                    break
+                yield chunk
+
+        async def _run_analysis() -> None:
+            """Run the ebur128 measurement in a background FFmpeg process."""
+            try:
+                async with FFMpeg(
+                    audio_input=_chunk_generator(),
+                    input_format=pcm_format,
+                    output_format=pcm_format,
+                    audio_output="NULL",
+                    filter_params=["ebur128=framelog=verbose"],
+                    collect_log_history=True,
+                    loglevel="info",
+                ) as ffmpeg_proc:
+                    await ffmpeg_proc.wait()
+                    log_lines_str = "\n".join(ffmpeg_proc.log_history)
+                    try:
+                        loudness_str = (
+                            log_lines_str.split("Integrated loudness")[1]
+                            .split("I:")[1]
+                            .split("LUFS")[0]
+                        )
+                        loudness = float(loudness_str.strip())
+                    except (IndexError, ValueError, AttributeError):
+                        self.logger.debug(
+                            "Could not determine loudness of %s from buffer analysis",
+                            streamdetails.uri,
+                        )
+                        return
+
+                    await self.mass.music.set_loudness(
+                        item_id, provider, loudness, media_type=media_type
+                    )
+                    self.logger.debug(
+                        "Buffer loudness measurement for %s: %s LUFS",
+                        streamdetails.uri,
+                        loudness,
+                    )
+            except Exception as err:
+                self.logger.debug("Loudness analysis from buffer failed: %s", err)
+
+        async def _attach() -> None:
+            # check if loudness already exists
+            if await self.mass.music.get_loudness(item_id, provider, media_type=media_type):
+                return
+            self.logger.debug("Attached loudness analyzer to buffer for %s", streamdetails.uri)
+            audio_buffer.register_chunk_callback(_on_chunk)
+            # start the FFmpeg analysis process
+            await _run_analysis()
+
+        self.mass.create_task(_attach)
+
+    def get_player_dsp_details(
+        self, player: Player, group_preventing_dsp: bool = False
+    ) -> DSPDetails:
+        """
+        Return DSP details of a single player.
+
+        :param player: The player to get DSP details for.
+        :param group_preventing_dsp: Whether grouping prevents DSP for this player.
+        """
+        dsp_config = self.mass.config.get_player_dsp_config(player.player_id)
+        dsp_state = DSPState.ENABLED if dsp_config.enabled else DSPState.DISABLED
+        if dsp_state == DSPState.ENABLED and (
+            group_preventing_dsp or is_grouping_preventing_dsp(player)
+        ):
+            dsp_state = DSPState.DISABLED_BY_UNSUPPORTED_GROUP
+            dsp_config = DSPConfig(enabled=False)
+        elif dsp_state == DSPState.DISABLED:
+            dsp_config = DSPConfig(enabled=False)
+
+        dsp_config.filters = [x for x in dsp_config.filters if x.enabled]
+
+        output_limiter = self.is_output_limiter_enabled(player)
+        return DSPDetails(
+            state=dsp_state,
+            input_gain=dsp_config.input_gain,
+            filters=dsp_config.filters,
+            output_gain=dsp_config.output_gain,
+            output_limiter=output_limiter,
+            output_format=player.extra_data.get("output_format", None),
+        )
+
+    def get_stream_dsp_details(self, queue_id: str) -> dict[str, DSPDetails]:
+        """Return DSP details of all players playing this queue, keyed by player_id."""
+        player = self.mass.players.get_player(queue_id)
+        dsp: dict[str, DSPDetails] = {}
+        assert player is not None
+        group_preventing_dsp = is_grouping_preventing_dsp(player)
+        output_format = None
+        is_external_group = False
+
+        if player.player_id.startswith(SGP_PREFIX):
+            if group_preventing_dsp:
+                sgp_player = cast("SyncGroupPlayer", player)
+                if sync_leader := sgp_player.sync_leader:
+                    output_format = sync_leader.extra_data.get("output_format", None)
+        else:
+            details = self.get_player_dsp_details(player)
+            dsp[player.player_id] = details
+            if group_preventing_dsp:
+                output_format = player.extra_data.get("output_format", None)
+            is_external_group = player.state.type in (PlayerType.GROUP, PlayerType.STEREO_PAIR)
+
+        if player and player.state.group_members and not is_external_group:
+            for child_id in player.state.group_members:
+                if child_id in dsp:
+                    continue
+                if child_player := self.mass.players.get_player(child_id):
+                    dsp[child_id] = self.get_player_dsp_details(
+                        child_player, group_preventing_dsp=group_preventing_dsp
+                    )
+                    if group_preventing_dsp:
+                        dsp[child_id].output_format = output_format
+        return dsp
+
+    def is_output_limiter_enabled(self, player: Player) -> bool:
+        """Check if the player has the output limiter enabled."""
+        deciding_player_id = player.player_id
+        if player.state.active_group:
+            deciding_player_id = player.state.active_group
+        elif player.state.synced_to:
+            deciding_player_id = player.state.synced_to
+        output_limiter_enabled = self.mass.config.get_raw_player_config_value(
+            deciding_player_id,
+            CONF_ENTRY_OUTPUT_LIMITER.key,
+            CONF_ENTRY_OUTPUT_LIMITER.default_value,
+        )
+        return bool(output_limiter_enabled)
+
+    def get_player_filter_params(
+        self, player_id: str, input_format: AudioFormat, output_format: AudioFormat
+    ) -> list[str]:
+        """Get player specific filter parameters for ffmpeg (if any)."""
+        filter_params = []
+
+        player = self.mass.players.get_player(player_id)
+        dsp_player_id = player_id
+        if player and player.protocol_parent_id:
+            dsp_player_id = player.protocol_parent_id
+        dsp = self.mass.config.get_player_dsp_config(dsp_player_id)
+        limiter_enabled = True
+
+        if player:
+            if is_grouping_preventing_dsp(player):
+                dsp.enabled = False
+            elif player.provider.domain == "player_group" and (
+                PlayerFeature.MULTI_DEVICE_DSP not in player.state.supported_features
+            ):
+                if player.state.group_members:
+                    child_player = self.mass.players.get_player(player.state.group_members[0])
+                    assert child_player is not None
+                    dsp = self.mass.config.get_player_dsp_config(child_player.player_id)
+                else:
+                    dsp.enabled = False
+
+            player.extra_data["output_format"] = output_format
+            if player.protocol_parent_id:
+                parent_player = self.mass.players.get_player(player.protocol_parent_id)
+                if parent_player:
+                    parent_player.extra_data["output_format"] = output_format
+
+            limiter_enabled = self.is_output_limiter_enabled(player)
+
+        if dsp.enabled:
+            if dsp.input_gain != 0:
+                filter_params.append(f"volume={dsp.input_gain}dB")
+            for f in dsp.filters:
+                if not f.enabled:
+                    continue
+                filter_params.extend(filter_to_ffmpeg_params(f, input_format))
+            if dsp.output_gain != 0:
+                filter_params.append(f"volume={dsp.output_gain}dB")
+
+        conf_channels = self.mass.config.get_raw_player_config_value(
+            player_id, CONF_OUTPUT_CHANNELS, "stereo"
+        )
+        if conf_channels == "left":
+            filter_params.append("pan=mono|c0=FL")
+        elif conf_channels == "right":
+            filter_params.append("pan=mono|c0=FR")
+
+        if limiter_enabled:
+            filter_params.append("alimiter=limit=-2dB:level=false:asc=true")
+
+        self.logger.debug("Generated ffmpeg params for player %s: %s", player_id, filter_params)
+        return filter_params
+
+    async def get_output_format(
+        self,
+        output_format_str: str,
+        player: Player,
+        content_sample_rate: int,
+        content_bit_depth: int,
+    ) -> AudioFormat:
+        """Parse (player specific) output format details for given format string."""
+        content_type: ContentType = ContentType.try_parse(output_format_str)
+        supported_rates_conf = cast(
+            "list[tuple[str, str]]",
+            await self.mass.config.get_player_config_value(
+                player.player_id, CONF_SAMPLE_RATES, unpack_splitted_values=True
+            ),
+        )
+        output_channels_str = self.mass.config.get_raw_player_config_value(
+            player.player_id, CONF_OUTPUT_CHANNELS, "stereo"
+        )
+        supported_sample_rates = tuple(int(x[0]) for x in supported_rates_conf)
+        supported_bit_depths = tuple(int(x[1]) for x in supported_rates_conf)
+
+        player_max_bit_depth = max(supported_bit_depths)
+        output_bit_depth = min(content_bit_depth, player_max_bit_depth)
+        if content_sample_rate in supported_sample_rates:
+            output_sample_rate = content_sample_rate
+        else:
+            output_sample_rate = max(supported_sample_rates)
+
+        if not content_type.is_lossless():
+            # no point in having a higher bit depth for lossy formats
+            output_bit_depth = 16
+            output_sample_rate = min(48000, output_sample_rate)
+        if output_format_str == "pcm":
+            content_type = ContentType.from_bit_depth(output_bit_depth)
+        fmt = AudioFormat(
+            content_type=content_type,
+            sample_rate=output_sample_rate,
+            bit_depth=output_bit_depth,
+            channels=1 if output_channels_str != "stereo" else 2,
+        )
+        fmt.bit_rate = get_bit_rate(fmt)
+        return fmt
+
+    async def select_flow_format(self, player: Player) -> AudioFormat:
+        """Parse (player specific) flow stream PCM format."""
+        supported_rates_conf = cast(
+            "list[tuple[str, str]]",
+            await self.mass.config.get_player_config_value(
+                player.player_id, CONF_SAMPLE_RATES, unpack_splitted_values=True
+            ),
+        )
+        supported_sample_rates = tuple(int(x[0]) for x in supported_rates_conf)
+        output_sample_rate = INTERNAL_PCM_FORMAT.sample_rate
+        for sample_rate in (192000, 96000, 48000, 44100):
+            if sample_rate in supported_sample_rates:
+                output_sample_rate = sample_rate
+                break
+        return AudioFormat(
+            content_type=INTERNAL_PCM_FORMAT.content_type,
+            sample_rate=output_sample_rate,
+            bit_depth=INTERNAL_PCM_FORMAT.bit_depth,
+            channels=2,
+        )
+
+    async def select_pcm_format(
+        self, player: Player, streamdetails: StreamDetails, smartfades_enabled: bool
+    ) -> AudioFormat:
+        """
+        Select the internal PCM format for streaming a queue item.
+
+        Uses F32 (float32) when audio processing is expected (volume normalization,
+        crossfade, DSP) for extra headroom. Falls back to the source material's
+        native bit depth when no processing is needed, avoiding unnecessary conversion.
+
+        :param player: The player requesting the stream.
+        :param streamdetails: Stream details for the current item.
+        :param smartfades_enabled: Whether crossfade is enabled for this stream.
+        """
+        supported_rates_conf = cast(
+            "list[tuple[str, str]]",
+            await self.mass.config.get_player_config_value(
+                player.player_id, CONF_SAMPLE_RATES, unpack_splitted_values=True
+            ),
+        )
+        supported_sample_rates = tuple(int(x[0]) for x in supported_rates_conf)
+        output_sample_rate = max(
+            (r for r in supported_sample_rates if r <= streamdetails.audio_format.sample_rate),
+            default=48000,
+        )
+
+        # determine if audio processing will be applied
+        # if so, use F32 for headroom; otherwise use source bit depth
+        needs_processing = (
+            smartfades_enabled
+            or streamdetails.volume_normalization_mode != VolumeNormalizationMode.DISABLED
+            or self.mass.config.get_player_dsp_config(player.player_id).enabled
+        )
+
+        if needs_processing:
+            content_type = INTERNAL_PCM_FORMAT.content_type
+            bit_depth = INTERNAL_PCM_FORMAT.bit_depth
+        else:
+            bit_depth = streamdetails.audio_format.bit_depth
+            content_type = ContentType.from_bit_depth(bit_depth)
+
+        pcm_format = AudioFormat(
+            sample_rate=output_sample_rate,
+            content_type=content_type,
+            bit_depth=bit_depth,
+            channels=streamdetails.audio_format.channels,
+        )
+        if smartfades_enabled:
+            pcm_format.channels = 2
+
+        return pcm_format
+
+    async def get_queue_item_stream(
+        self,
+        queue_item: QueueItem,
+        pcm_format: AudioFormat,
+        seek_position: int = 0,
+        playback_speed: float = 1.0,
+        raise_on_error: bool = True,
+    ) -> AsyncGenerator[bytes, None]:
+        """
+        Get the (PCM) audio stream for a single queue item.
+
+        Audio is always served from the AudioBuffer which stores raw decoded PCM.
+        Volume normalization and other filters are applied on-the-fly when reading
+        from the buffer.
+        """
+        streamdetails = queue_item.streamdetails
+        assert streamdetails
+        filter_params: list[str] = []
+
+        # handle volume normalization
+        gain_correct: float | None = None
+        if streamdetails.volume_normalization_mode == VolumeNormalizationMode.DYNAMIC:
+            filter_rule = (
+                f"loudnorm=I={streamdetails.target_loudness}"
+                ":TP=-2.0:LRA=10.0:offset=0.0:print_format=json"
+            )
+            filter_params.append(filter_rule)
+        elif streamdetails.volume_normalization_mode == VolumeNormalizationMode.FIXED_GAIN:
+            config_key = (
+                CONF_VOLUME_NORMALIZATION_FIXED_GAIN_TRACKS
+                if streamdetails.media_type == MediaType.TRACK
+                else CONF_VOLUME_NORMALIZATION_FIXED_GAIN_RADIO
+            )
+            gain_value = await self.mass.config.get_core_config_value(
+                "streams", config_key, default=0.0, return_type=float
+            )
+            gain_correct = round(gain_value, 2)
+            filter_params.append(f"volume={gain_correct}dB")
+        elif streamdetails.volume_normalization_mode == VolumeNormalizationMode.MEASUREMENT_ONLY:
+            target_loudness = (
+                float(streamdetails.target_loudness)
+                if streamdetails.target_loudness is not None
+                else 0.0
+            )
+            if streamdetails.prefer_album_loudness and streamdetails.loudness_album is not None:
+                gain_correct = target_loudness - float(streamdetails.loudness_album)
+            elif streamdetails.loudness is not None:
+                gain_correct = target_loudness - float(streamdetails.loudness)
+            else:
+                gain_correct = 0.0
+            gain_correct = round(gain_correct, 2)
+            filter_params.append(f"volume={gain_correct}dB")
+        streamdetails.volume_normalization_gain_correct = gain_correct
+
+        # handle playback speed
+        if playback_speed != 1.0:
+            filter_params.append(f"atempo={playback_speed}")
+
+        # handle optional fade-in
+        if streamdetails.fade_in:
+            filter_params.insert(0, "afade=type=in:start_time=0:duration=3")
+
+        self.logger.debug(
+            "Starting queue item stream for %s (%s)"
+            " - using fade-in: %s"
+            " - using volume normalization: %s",
+            queue_item.name,
+            streamdetails.uri,
+            streamdetails.fade_in,
+            streamdetails.volume_normalization_mode,
+        )
+
+        # get or create the AudioBuffer (stores raw decoded PCM)
+        seek_position_ms = int(seek_position * 1000)
+        audio_buffer = await AudioBuffer.get_buffer(
+            mass=self.mass, streamdetails=streamdetails, seek_position_ms=seek_position_ms
+        )
+
+        # read from buffer with filters applied (volume normalization, speed, fade-in, etc.)
+        # if no processing needed, this yields directly from the buffer
+        media_stream_gen = audio_buffer.get_stream(
+            output_format=pcm_format,
+            seek_position_ms=seek_position_ms,
+            filter_params=filter_params if filter_params else None,
+        )
+
+        first_chunk_received = False
+        bytes_received = 0
+        finished = False
+        stream_started_at = asyncio.get_event_loop().time()
+        try:
+            async for chunk in media_stream_gen:
+                bytes_received += len(chunk)
+                if not first_chunk_received:
+                    first_chunk_received = True
+                    self.logger.debug(
+                        "First audio chunk received for %s (%s) after %.2f seconds",
+                        queue_item.name,
+                        streamdetails.uri,
+                        asyncio.get_event_loop().time() - stream_started_at,
+                    )
+                yield chunk
+                del chunk
+            finished = True
+        except AudioError as err:
+            streamdetails.stream_error = True
+            queue_item.available = False
+            if raise_on_error:
+                raise
+            self.logger.error(
+                "AudioError while streaming queue item %s (%s): %s",
+                queue_item.name,
+                streamdetails.uri,
+                err,
+            )
+        except asyncio.CancelledError:
+            raise
+        except Exception as err:
+            streamdetails.stream_error = True
+            if raise_on_error:
+                raise
+            self.logger.exception(
+                "Unexpected error while streaming queue item %s (%s): %s",
+                queue_item.name,
+                streamdetails.uri,
+                err,
+            )
+        finally:
+            seconds_streamed = bytes_received / pcm_format.pcm_sample_size
+            streamdetails.seconds_streamed = seconds_streamed
+            self.logger.debug(
+                "stream %s for %s in %.2f seconds - seconds streamed/buffered: %.2f",
+                "aborted" if not finished else "finished",
+                streamdetails.uri,
+                asyncio.get_event_loop().time() - stream_started_at,
+                seconds_streamed,
+            )
+            if (finished or seconds_streamed >= 90) and (
+                music_prov := self.mass.get_provider(streamdetails.provider)
+            ):
+                if TYPE_CHECKING:
+                    assert isinstance(music_prov, MusicProvider)
+                self.mass.create_task(music_prov.on_streamed(streamdetails))
+
+    async def get_queue_item_stream_with_smartfade(
+        self,
+        player: Player,
+        queue_item: QueueItem,
+        pcm_format: AudioFormat,
+        smart_fades_mode: SmartFadesMode = SmartFadesMode.SMART_CROSSFADE,
+        standard_crossfade_duration: int = 10,
+    ) -> AsyncGenerator[bytes, None]:
+        """Get the audio stream for a single queue item with (smart) crossfade to the next item."""
+        queue = self.mass.player_queues.get(queue_item.queue_id)
+        if not queue:
+            raise RuntimeError(f"Queue {queue_item.queue_id} not found")
+
+        streamdetails = queue_item.streamdetails
+        assert streamdetails
+        crossfade_data = self._crossfade_data.pop(queue.queue_id, None)
+
+        if crossfade_data and streamdetails.seek_position > 0:
+            # don't do crossfade when seeking into track
+            crossfade_data = None
+        if crossfade_data and (crossfade_data.queue_item_id != queue_item.queue_item_id):
+            # edge case alert: the next item changed just while we were preloading/crossfading
+            self.logger.warning(
+                "Skipping crossfade data for queue %s - next item changed!", queue.display_name
+            )
+            crossfade_data = None
+
+        self.logger.debug(
+            "Start Streaming queue track: %s (%s) for queue %s "
+            "- crossfade mode: %s "
+            "- crossfading from previous track: %s ",
+            queue_item.streamdetails.uri if queue_item.streamdetails else "Unknown URI",
+            queue_item.name,
+            queue.display_name,
+            smart_fades_mode,
+            "true" if crossfade_data else "false",
+        )
+
+        buffer = b""
+        bytes_written = 0
+        # calculate crossfade buffer size
+        crossfade_buffer_duration = (
+            SMART_CROSSFADE_DURATION
+            if smart_fades_mode == SmartFadesMode.SMART_CROSSFADE
+            else standard_crossfade_duration
+        )
+        crossfade_buffer_duration = min(
+            crossfade_buffer_duration,
+            int(streamdetails.duration / 2)
+            if streamdetails.duration
+            else crossfade_buffer_duration,
+        )
+        # Ensure crossfade buffer size is aligned to frame boundaries
+        # Frame size = bytes_per_sample * channels
+        bytes_per_sample = pcm_format.bit_depth // 8
+        frame_size = bytes_per_sample * pcm_format.channels
+        crossfade_buffer_size = int(pcm_format.pcm_sample_size * crossfade_buffer_duration)
+        # Round down to nearest frame boundary
+        crossfade_buffer_size = (crossfade_buffer_size // frame_size) * frame_size
+        fade_out_data: bytes | None = None
+
+        if crossfade_data:
+            # Calculate discard amount in seconds (format-independent)
+            # Use fade_in_pcm_format because fade_in_size is in the next track's original format
+            fade_in_duration_seconds = (
+                crossfade_data.fade_in_size / crossfade_data.fade_in_pcm_format.pcm_sample_size
+            )
+            discard_seconds = int(fade_in_duration_seconds) - 1
+            # Calculate discard amounts in CURRENT track's format
+            discard_bytes = int(discard_seconds * pcm_format.pcm_sample_size)
+            # Convert fade_in_size to current track's format for correct leftover calculation
+            fade_in_size_in_current_format = int(
+                fade_in_duration_seconds * pcm_format.pcm_sample_size
+            )
+            discard_leftover = fade_in_size_in_current_format - discard_bytes
+        else:
+            discard_seconds = streamdetails.seek_position
+            discard_leftover = 0
+        total_chunks_received = 0
+        req_buffer_size = crossfade_buffer_size
+        async for chunk in self.get_queue_item_stream(
+            queue_item,
+            pcm_format,
+            seek_position=discard_seconds,
+            playback_speed=cast("float", queue_item.extra_attributes.get("playback_speed", 1.0)),
+        ):
+            total_chunks_received += 1
+            if discard_leftover:
+                # discard leftover bytes from crossfade data
+                chunk = chunk[discard_leftover:]  # noqa: PLW2901
+                discard_leftover = 0
+
+            if total_chunks_received < 10:
+                # we want a stream to start as quickly as possible
+                # so for the first 10 chunks we keep a very short buffer
+                req_buffer_size = pcm_format.pcm_sample_size
+            else:
+                req_buffer_size = crossfade_buffer_size
+
+            # ALWAYS APPEND CHUNK TO BUFFER
+            buffer += chunk
+            del chunk
+            if len(buffer) < req_buffer_size:
+                # buffer is not full enough, move on
+                continue
+
+            ####  HANDLE CROSSFADE DATA FROM PREVIOUS TRACK
+            if crossfade_data:
+                # send the (second half of the) crossfade data
+                if crossfade_data.pcm_format != pcm_format:
+                    # edge case: pcm format mismatch, we need to resample
+                    self.logger.debug(
+                        "Resampling crossfade data from %s to %s for queue %s",
+                        crossfade_data.pcm_format.sample_rate,
+                        pcm_format.sample_rate,
+                        queue.display_name,
+                    )
+                    resampled_data = await resample_pcm_audio(
+                        crossfade_data.data, crossfade_data.pcm_format, pcm_format
+                    )
+                    if resampled_data:
+                        for _chunk in divide_chunks(resampled_data, pcm_format.pcm_sample_size):
+                            yield _chunk
+                        bytes_written += len(resampled_data)
+                    else:
+                        # Resampling failed, error already logged in resample_pcm_audio
+                        # Skip crossfade data entirely - stream continues without it
+                        self.logger.warning(
+                            "Skipping crossfade data for queue %s due to resampling failure",
+                            queue.display_name,
+                        )
+                else:
+                    for _chunk in divide_chunks(crossfade_data.data, pcm_format.pcm_sample_size):
+                        yield _chunk
+                    bytes_written += len(crossfade_data.data)
+                # clear vars
+                crossfade_data = None
+
+            #### OTHER: enough data in buffer, feed to output
+            while len(buffer) > req_buffer_size:
+                yield buffer[: pcm_format.pcm_sample_size]
+                bytes_written += pcm_format.pcm_sample_size
+                buffer = buffer[pcm_format.pcm_sample_size :]
+
+        #### HANDLE END OF TRACK
+
+        if crossfade_data:
+            # edge case: we did not get enough data to send the crossfade data
+            # send the (second half of the) crossfade data
+            if crossfade_data.pcm_format != pcm_format:
+                # (yet another) edge case: pcm format mismatch, we need to resample
+                self.logger.debug(
+                    "Resampling remaining crossfade data from %s to %s for queue %s",
+                    crossfade_data.pcm_format.sample_rate,
+                    pcm_format.sample_rate,
+                    queue.display_name,
+                )
+                resampled_crossfade_data = await resample_pcm_audio(
+                    crossfade_data.data, crossfade_data.pcm_format, pcm_format
+                )
+                if resampled_crossfade_data:
+                    crossfade_data.data = resampled_crossfade_data
+                else:
+                    # Resampling failed, error already logged in resample_pcm_audio
+                    # Skip the crossfade data entirely
+                    self.logger.warning(
+                        "Skipping remaining crossfade data for queue %s due to resampling failure",
+                        queue.display_name,
+                    )
+                    crossfade_data = None
+            if crossfade_data:
+                for _chunk in divide_chunks(crossfade_data.data, pcm_format.pcm_sample_size):
+                    yield _chunk
+                bytes_written += len(crossfade_data.data)
+                crossfade_data = None
+
+        # get next track for crossfade
+        next_queue_item: QueueItem | None
+        try:
+            self.logger.debug(
+                "Preloading NEXT track for crossfade for queue %s", queue.display_name
+            )
+            next_queue_item = await self.mass.player_queues.load_next_queue_item(
+                queue.queue_id, queue_item.queue_item_id
+            )
+            # set index_in_buffer to prevent our next track is overwritten while preloading
+            if next_queue_item.streamdetails is None:
+                raise InvalidDataError(
+                    f"No streamdetails for next queue item {next_queue_item.queue_item_id}"
+                )
+            queue.index_in_buffer = self.mass.player_queues.index_by_id(
+                queue.queue_id, next_queue_item.queue_item_id
+            )
+            next_queue_item_pcm_format = await self.select_pcm_format(
+                player=player, streamdetails=next_queue_item.streamdetails, smartfades_enabled=True
+            )
+        except QueueEmpty:
+            # end of queue reached, no next item
+            next_queue_item = None
+
+        if not next_queue_item or not self.crossfade_allowed(
+            queue_item,
+            smart_fades_mode=smart_fades_mode,
+            flow_mode=False,
+            next_queue_item=next_queue_item,
+            sample_rate=pcm_format.sample_rate,
+            next_sample_rate=next_queue_item_pcm_format.sample_rate,
+        ):
+            # no crossfade enabled/allowed, just yield the buffer last part
+            bytes_written += len(buffer)
+            for _chunk in divide_chunks(buffer, pcm_format.pcm_sample_size):
+                yield _chunk
+        else:
+            # if crossfade is enabled, save fadeout part in buffer to pickup for next track
+            fade_out_data = buffer
+            buffer = b""
+            try:
+                async for chunk in self.get_queue_item_stream(
+                    next_queue_item,
+                    next_queue_item_pcm_format,
+                    playback_speed=cast(
+                        "float", queue_item.extra_attributes.get("playback_speed", 1.0)
+                    ),
+                ):
+                    # append to buffer until we reach crossfade size
+                    # we only need the first X seconds of the NEXT track so we can
+                    # perform the crossfade.
+                    # the crossfaded audio of the previous and next track will be
+                    # sent in two equal parts: first half now, second half
+                    # when the next track starts. We use CrossfadeData to store
+                    # the second half to be picked up by the next track's stream generator.
+                    # Note that we more or less expect the user to have enabled the in-memory
+                    # buffer so we can keep the next track's audio data in memory.
+                    buffer += chunk
+                    del chunk
+                    if len(buffer) >= crossfade_buffer_size:
+                        break
+                ####  HANDLE CROSSFADE OF PREVIOUS TRACK AND NEW TRACK
+                # Store original buffer size before any resampling for fade_in_size calculation
+                # This size is in the next track's original format which is what we need
+                original_buffer_size = len(buffer)
+                if next_queue_item_pcm_format != pcm_format:
+                    # edge case: pcm format mismatch, we need to resample the next track's
+                    # beginning part before crossfading
+                    self.logger.debug(
+                        "Resampling next track's crossfade from %s to %s for queue %s",
+                        next_queue_item_pcm_format.sample_rate,
+                        pcm_format.sample_rate,
+                        queue.display_name,
+                    )
+                    buffer = await resample_pcm_audio(
+                        buffer, next_queue_item_pcm_format, pcm_format
+                    )
+                # perform actual (smart fades) crossfade using mixer
+                crossfade_bytes = await self._smart_fades_mixer.mix(
+                    fade_in_part=buffer,
+                    fade_out_part=fade_out_data,
+                    fade_in_streamdetails=cast("StreamDetails", next_queue_item.streamdetails),
+                    fade_out_streamdetails=streamdetails,
+                    pcm_format=pcm_format,
+                    standard_crossfade_duration=standard_crossfade_duration,
+                    mode=smart_fades_mode,
+                )
+                # send half of the crossfade_part (= approx the fadeout part)
+                split_point = (len(crossfade_bytes) + 1) // 2
+                crossfade_first = crossfade_bytes[:split_point]
+                crossfade_second = crossfade_bytes[split_point:]
+                del crossfade_bytes
+                bytes_written += len(crossfade_first)
+                for _chunk in divide_chunks(crossfade_first, pcm_format.pcm_sample_size):
+                    yield _chunk
+                # store the other half for the next track
+                # IMPORTANT: crossfade_second data is in CURRENT track's format (pcm_format)
+                # because it was created from the resampled buffer used for mixing.
+                # BUT fade_in_size represents bytes in NEXT track's original format
+                # (next_queue_item_pcm_format) because that's how much of the next track
+                # was consumed during the crossfade. We need both formats to correctly
+                # handle the crossfade data when the next track starts.
+                self._crossfade_data[queue_item.queue_id] = CrossfadeData(
+                    data=crossfade_second,
+                    fade_in_size=original_buffer_size,
+                    pcm_format=pcm_format,  # Format of the data (current track)
+                    fade_in_pcm_format=next_queue_item_pcm_format,  # Format for fade_in_size
+                    queue_item_id=next_queue_item.queue_item_id,
+                )
+            except AudioError:
+                # no crossfade possible, just yield the fade_out_data
+                next_queue_item = None
+                yield fade_out_data
+                bytes_written += len(fade_out_data)
+                del fade_out_data
+        # make sure the buffer gets cleaned up
+        del buffer
+        # update duration details based on the actual pcm data we sent
+        # this also accounts for crossfade and silence stripping
+        seconds_streamed = bytes_written / pcm_format.pcm_sample_size
+        streamdetails.seconds_streamed = seconds_streamed
+        streamdetails.duration = int(streamdetails.seek_position + seconds_streamed)
+        self.logger.debug(
+            "Finished Streaming queue track: %s (%s) on queue %s "
+            "- crossfade data prepared for next track: %s",
+            streamdetails.uri,
+            queue_item.name,
+            queue.display_name,
+            next_queue_item.name if next_queue_item else "N/A",
+        )
+
+    async def get_queue_flow_stream(
+        self, queue: PlayerQueue, start_queue_item: QueueItem, pcm_format: AudioFormat
+    ) -> AsyncGenerator[bytes, None]:
+        """
+        Get a flow stream of all tracks in the queue as raw PCM audio.
+
+        yields chunks of exactly 1 second of audio in the given pcm_format.
+        """
+        # ruff: noqa: PLR0915
+        assert pcm_format.content_type.is_pcm()
+        queue_track = None
+        last_fadeout_part: bytes = b""
+        last_streamdetails: StreamDetails | None = None
+        last_play_log_entry: PlayLogEntry | None = None
+        queue.flow_mode = True
+        queue.flow_mode_stream_log = []
+        if not start_queue_item:
+            # this can happen in some (edge case) race conditions
+            return
+        pcm_sample_size = pcm_format.pcm_sample_size
+        if start_queue_item.media_type != MediaType.TRACK:
+            # no crossfade on non-tracks
+            smart_fades_mode = SmartFadesMode.DISABLED
+            standard_crossfade_duration = 0
+        else:
+            smart_fades_mode = await self.mass.config.get_player_config_value(
+                queue.queue_id, CONF_SMART_FADES_MODE, return_type=SmartFadesMode
+            )
+            standard_crossfade_duration = self.mass.config.get_raw_player_config_value(
+                queue.queue_id, CONF_CROSSFADE_DURATION, 10
+            )
+        # smart crossfade requires a large buffer for beat analysis
+        if (
+            smart_fades_mode == SmartFadesMode.SMART_CROSSFADE
+            and self.mass.config.get_raw_core_config_value(
+                "streams", CONF_BUFFER_SIZE, CONF_BUFFER_SIZE_DEFAULT
+            )
+            == BufferSize.MINIMAL
+        ):
+            smart_fades_mode = SmartFadesMode.STANDARD_CROSSFADE
+        self.logger.info(
+            "Start Queue Flow stream for Queue %s - crossfade: %s %s",
+            queue.display_name,
+            smart_fades_mode,
+            f"({standard_crossfade_duration}s)"
+            if smart_fades_mode == SmartFadesMode.STANDARD_CROSSFADE
+            else "",
+        )
+        total_bytes_sent = 0
+        total_chunks_received = 0
+
+        while True:
+            # get (next) queue item to stream
+            if queue_track is None:
+                queue_track = start_queue_item
+            else:
+                try:
+                    queue_track = await self.mass.player_queues.load_next_queue_item(
+                        queue.queue_id, queue_track.queue_item_id
+                    )
+                except QueueEmpty:
+                    break
+
+            if queue_track.streamdetails is None:
+                self.logger.error(
+                    "No StreamDetails for queue item %s (%s) on queue %s - skipping track",
+                    queue_track.queue_item_id,
+                    queue_track.name,
+                    queue.display_name,
+                )
+                continue
+
+            self.logger.debug(
+                "Start Streaming queue track: %s (%s) for queue %s",
+                queue_track.streamdetails.uri,
+                queue_track.name,
+                queue.display_name,
+            )
+            # append to play log so the queue controller can work out which track is playing
+            play_log_entry = PlayLogEntry(queue_track.queue_item_id)
+            queue.flow_mode_stream_log.append(play_log_entry)
+            # calculate crossfade buffer size
+            crossfade_buffer_duration = (
+                SMART_CROSSFADE_DURATION
+                if smart_fades_mode == SmartFadesMode.SMART_CROSSFADE
+                else standard_crossfade_duration
+            )
+            crossfade_buffer_duration = min(
+                crossfade_buffer_duration,
+                int(queue_track.streamdetails.duration / 2)
+                if queue_track.streamdetails.duration
+                else crossfade_buffer_duration,
+            )
+            # Ensure crossfade buffer size is aligned to frame boundaries
+            # Frame size = bytes_per_sample * channels
+            bytes_per_sample = pcm_format.bit_depth // 8
+            frame_size = bytes_per_sample * pcm_format.channels
+            crossfade_buffer_size = int(pcm_format.pcm_sample_size * crossfade_buffer_duration)
+            # Round down to nearest frame boundary
+            crossfade_buffer_size = (crossfade_buffer_size // frame_size) * frame_size
+
+            bytes_written = 0
+            buffer = b""
+            # handle incoming audio chunks
+            first_chunk_received = False
+            # buffer size needs to be big enough to include the crossfade part
+
+            async for chunk in self.get_queue_item_stream(
+                queue_track,
+                pcm_format=pcm_format,
+                seek_position=queue_track.streamdetails.seek_position,
+                playback_speed=cast(
+                    "float", queue_track.extra_attributes.get("playback_speed", 1.0)
+                ),
+                raise_on_error=False,
+            ):
+                total_chunks_received += 1
+                if not first_chunk_received:
+                    first_chunk_received = True
+                    # inform the queue that the track is now loaded in the buffer
+                    # so the next track can be preloaded
+                    self.mass.player_queues.track_loaded_in_buffer(
+                        queue.queue_id, queue_track.queue_item_id
+                    )
+                if total_chunks_received < 10 and smart_fades_mode != SmartFadesMode.DISABLED:
+                    # we want a stream to start as quickly as possible
+                    # so for the first 10 chunks we keep a very short buffer
+                    req_buffer_size = pcm_format.pcm_sample_size
+                else:
+                    req_buffer_size = (
+                        pcm_sample_size
+                        if smart_fades_mode == SmartFadesMode.DISABLED
+                        else crossfade_buffer_size
+                    )
+
+                # ALWAYS APPEND CHUNK TO BUFFER
+                buffer += chunk
+                del chunk
+                if len(buffer) < req_buffer_size:
+                    # buffer is not full enough, move on
+                    # yield control to event loop with 10ms delay
+                    await asyncio.sleep(0.01)
+                    continue
+
+                ####  HANDLE CROSSFADE OF PREVIOUS TRACK AND NEW TRACK
+                if last_fadeout_part and last_streamdetails:
+                    # perform crossfade
+                    fadein_part = buffer[:crossfade_buffer_size]
+                    remaining_bytes = buffer[crossfade_buffer_size:]
+                    # Use the mixer to handle all crossfade logic
+                    try:
+                        crossfade_part = await self._smart_fades_mixer.mix(
+                            fade_in_part=fadein_part,
+                            fade_out_part=last_fadeout_part,
+                            fade_in_streamdetails=queue_track.streamdetails,
+                            fade_out_streamdetails=last_streamdetails,
+                            pcm_format=pcm_format,
+                            standard_crossfade_duration=standard_crossfade_duration,
+                            mode=smart_fades_mode,
+                        )
+                    except Exception as mix_err:
+                        self.logger.warning(
+                            "Crossfade mixer failed for %s, falling back to simple concat: %s",
+                            queue_track.name,
+                            mix_err,
+                        )
+                        # Fallback: just output the fadeout part then the buffer
+                        for _chunk in divide_chunks(last_fadeout_part, pcm_sample_size):
+                            yield _chunk
+                            bytes_written += len(_chunk)
+                            del _chunk
+                        crossfade_part = b""
+                        remaining_bytes = buffer
+                    if crossfade_part:
+                        # because the crossfade exists of both the fadein and fadeout part
+                        # we need to correct the bytes_written accordingly so the duration
+                        # calculations at the end of the track are correct
+                        crossfade_part_len = len(crossfade_part)
+                        bytes_written += int(crossfade_part_len / 2)
+                        if last_play_log_entry:
+                            assert last_play_log_entry.seconds_streamed is not None
+                            last_play_log_entry.seconds_streamed += (
+                                crossfade_part_len / 2 / pcm_sample_size
+                            )
+                        # yield crossfade_part (in pcm_sample_size chunks)
+                        for _chunk in divide_chunks(crossfade_part, pcm_sample_size):
+                            yield _chunk
+                            del _chunk
+                        del crossfade_part
+                    # also write the leftover bytes from the crossfade action
+                    if remaining_bytes:
+                        yield remaining_bytes
+                        bytes_written += len(remaining_bytes)
+                        del remaining_bytes
+                    # clear vars
+                    last_fadeout_part = b""
+                    last_streamdetails = None
+                    buffer = b""
+
+                #### OTHER: enough data in buffer, feed to output
+                while len(buffer) > req_buffer_size:
+                    yield buffer[:pcm_sample_size]
+                    bytes_written += pcm_sample_size
+                    buffer = buffer[pcm_sample_size:]
+
+            #### HANDLE END OF TRACK
+            if not first_chunk_received:
+                # Track failed to stream - no chunks received at all
+                self.logger.warning(
+                    "Track %s (%s) on queue %s produced no audio data - skipping",
+                    queue_track.name,
+                    queue_track.streamdetails.uri if queue_track.streamdetails else "unknown",
+                    queue.display_name,
+                )
+                # Clean up and continue to next track
+                queue_track.streamdetails.stream_error = True
+                play_log_entry.seconds_streamed = 0
+                del buffer
+                continue
+            if last_fadeout_part:
+                # edge case: we did not get enough data to make the crossfade
+                for _chunk in divide_chunks(last_fadeout_part, pcm_sample_size):
+                    yield _chunk
+                    del _chunk
+                bytes_written += len(last_fadeout_part)
+                last_fadeout_part = b""
+            if self.crossfade_allowed(
+                queue_track, smart_fades_mode=smart_fades_mode, flow_mode=True
+            ):
+                # if crossfade is enabled, save fadeout part to pickup for next track
+                last_fadeout_part = buffer[-crossfade_buffer_size:]
+                last_streamdetails = queue_track.streamdetails
+                last_play_log_entry = play_log_entry
+                remaining_bytes = buffer[:-crossfade_buffer_size]
+                if remaining_bytes:
+                    yield remaining_bytes
+                    bytes_written += len(remaining_bytes)
+                del remaining_bytes
+            elif buffer:
+                # no crossfade enabled, just yield the buffer last part
+                bytes_written += len(buffer)
+                for _chunk in divide_chunks(buffer, pcm_sample_size):
+                    yield _chunk
+                    del _chunk
+            # make sure the buffer gets cleaned up
+            del buffer
+
+            # update duration details based on the actual pcm data we sent
+            # this also accounts for crossfade and silence stripping
+            seconds_streamed = bytes_written / pcm_sample_size
+            queue_track.streamdetails.seconds_streamed = seconds_streamed
+            queue_track.streamdetails.duration = int(
+                queue_track.streamdetails.seek_position + seconds_streamed
+            )
+            play_log_entry.seconds_streamed = seconds_streamed
+            play_log_entry.duration = queue_track.streamdetails.duration
+            total_bytes_sent += bytes_written
+            self.logger.debug(
+                "Finished Streaming queue track: %s (%s) on queue %s",
+                queue_track.streamdetails.uri,
+                queue_track.name,
+                queue.display_name,
+            )
+        #### HANDLE END OF QUEUE FLOW STREAM
+        # end of queue flow: make sure we yield the last_fadeout_part
+        if last_fadeout_part:
+            for _chunk in divide_chunks(last_fadeout_part, pcm_sample_size):
+                yield _chunk
+                del _chunk
+            # correct seconds streamed/duration
+            last_part_seconds = len(last_fadeout_part) / pcm_sample_size
+            streamdetails = queue_track.streamdetails
+            assert streamdetails is not None
+            streamdetails.seconds_streamed = (
+                streamdetails.seconds_streamed or 0
+            ) + last_part_seconds
+            streamdetails.duration = int((streamdetails.duration or 0) + last_part_seconds)
+            last_fadeout_part = b""
+        total_bytes_sent += bytes_written
+        self.logger.info("Finished Queue Flow stream for Queue %s", queue.display_name)
+        # inform the queue controller that all audio data has been generated
+        # so it can handle the case where new items were added after the flow stream ended
+        self.mass.player_queues.queue_buffer_completed(queue.queue_id)
+
+    def crossfade_allowed(
+        self,
+        queue_item: QueueItem,
+        smart_fades_mode: SmartFadesMode,
+        flow_mode: bool = False,
+        next_queue_item: QueueItem | None = None,
+        sample_rate: int | None = None,
+        next_sample_rate: int | None = None,
+    ) -> bool:
+        """Get the crossfade config for a queue item."""
+        if smart_fades_mode == SmartFadesMode.DISABLED:
+            return False
+        if not (self.mass.players.get_player(queue_item.queue_id)):
+            return False  # just a guard
+        if queue_item.media_type != MediaType.TRACK:
+            self.logger.debug("Skipping crossfade: current item is not a track")
+            return False
+        # check if the next item is part of the same album
+        next_item = next_queue_item or self.mass.player_queues.get_next_item(
+            queue_item.queue_id, queue_item.queue_item_id
+        )
+        if not next_item:
+            # there is no next item!
+            return False
+        # check if next item is a track
+        if next_item.media_type != MediaType.TRACK:
+            self.logger.debug("Skipping crossfade: next item is not a track")
+            return False
+        if (
+            isinstance(queue_item.media_item, Track)
+            and isinstance(next_item.media_item, Track)
+            and queue_item.media_item.album
+            and next_item.media_item.album
+            and queue_item.media_item.album == next_item.media_item.album
+            and not self.mass.config.get_raw_core_config_value(
+                "streams", CONF_ALLOW_CROSSFADE_SAME_ALBUM, False
+            )
+        ):
+            # in general, crossfade is not desired for tracks of the same (gapless) album
+            # because we have no accurate way to determine if the album is gapless or not,
+            # for now we just never crossfade between tracks of the same album
+            self.logger.debug("Skipping crossfade: next item is part of the same album")
+            return False
+
+        # check if we're allowed to crossfade on different sample rates
+        if (
+            not flow_mode
+            and sample_rate
+            and next_sample_rate
+            and sample_rate != next_sample_rate
+            and not self.mass.config.get_raw_player_config_value(
+                queue_item.queue_id,
+                CONF_ENTRY_SUPPORT_GAPLESS_DIFFERENT_SAMPLE_RATES.key,
+                CONF_ENTRY_SUPPORT_GAPLESS_DIFFERENT_SAMPLE_RATES.default_value,
+            )
+        ):
+            self.logger.debug(
+                "Skipping crossfade: player does not support gapless playback "
+                "with different sample rates (%s vs %s)",
+                sample_rate,
+                next_sample_rate,
+            )
+            return False
+
+        return True
+
+    def clear_crossfade_data(self, queue_id: str) -> None:
+        """
+        Clear any pending crossfade data for a queue.
+
+        :param queue_id: The queue ID to clear crossfade data for.
+        """
+        if queue_id in self._crossfade_data:
+            self.logger.debug("Clearing crossfade data for queue %s", queue_id)
+            del self._crossfade_data[queue_id]
+
+    # --- Private methods ---
+
+    @asynccontextmanager
+    async def _connect_radio_stream(self, url: str, **kwargs: Any) -> AsyncGenerator[Any, None]:
+        """
+        Connect to a radio stream URL with fallback for legacy SSL/TLS configurations.
+
+        Some radio servers use outdated TLS configurations that reject modern
+        cipher suites. Since radio streams are public broadcast content,
+        relaxing cipher requirements is acceptable.
+
+        :param url: The radio stream URL to connect to.
+        :param kwargs: Additional keyword arguments passed to aiohttp get().
+        """
+        try:
+            async with self.mass.http_session_no_ssl.get(url, **kwargs) as resp:
+                yield resp
+        except ClientConnectorSSLError:
+            self.logger.info(
+                "SSL handshake failed for %s, retrying with permissive cipher configuration", url
+            )
+            insecure_ssl_context = ssl_util.client_context_no_verify(
+                ssl_util.SSLCipherList.INSECURE
+            )
+            async with self.mass.http_session_no_ssl.get(
+                url, ssl=insecure_ssl_context, **kwargs
+            ) as resp:
+                yield resp
+
+    async def _update_hls_radio_metadata(
+        self,
+        streamdetails: StreamDetails,
+        elapsed_time: int,
+    ) -> None:
+        """
+        Update HLS radio stream metadata by fetching the playlist.
+
+        Fetches the HLS playlist and extracts metadata from EXTINF lines.
+
+        :param streamdetails: StreamDetails object to update with metadata
+        :param elapsed_time: Current playback position in seconds (unused for live radio)
+        """
+        mass = self.mass
+        try:
+            # Get the actual media playlist URL from cache or resolve it
+            # We cache the media_playlist_url in streamdetails.data to avoid re-resolving
+            if streamdetails.data is None:
+                streamdetails.data = {}
+            media_playlist_url = streamdetails.data.get("hls_media_playlist_url")
+            if not media_playlist_url:
+                try:
+                    assert isinstance(streamdetails.path, str)  # for type checking
+                    substream = await self.get_hls_substream(streamdetails.path)
+                    media_playlist_url = substream.path
+                    streamdetails.data["hls_media_playlist_url"] = media_playlist_url
+                except Exception as err:
+                    self.logger.warning(
+                        "Failed to resolve HLS substream for metadata monitoring: %s", err
+                    )
+                    return
+
+            # Fetch the media playlist
+            timeout = ClientTimeout(total=0, connect=10, sock_read=30)
+            async with mass.http_session_no_ssl.get(media_playlist_url, timeout=timeout) as resp:
+                resp.raise_for_status()
+                playlist_content = await resp.text()
+
+            # Parse the playlist and look for EXTINF metadata
+            # The most recent segment usually has the current metadata
+            lines = playlist_content.strip().split("\n")
+            for line in reversed(lines):
+                if line.startswith("#EXTINF:"):
+                    # Extract metadata from EXTINF line
+                    metadata = parse_extinf_metadata(line)
+
+                    # Build stream title from title and artist
+                    title = metadata.get("title", "")
+                    artist = metadata.get("artist", "")
+
+                    if title or artist:
+                        # Format as "Artist - Title"
+                        if artist and title:
+                            stream_title = f"{artist} - {title}"
+                        elif title:
+                            stream_title = title
+                        else:
+                            stream_title = artist
+
+                        # Clean the stream title
+                        cleaned_title = clean_stream_title(stream_title)
+
+                        # Only update if changed
+                        if cleaned_title != streamdetails.stream_title and cleaned_title:
+                            self.logger.log(
+                                VERBOSE_LOG_LEVEL, "HLS Radio metadata updated: %s", cleaned_title
+                            )
+                            streamdetails.stream_title = cleaned_title
+
+                    # Only check the most recent EXTINF
+                    break
+
+        except Exception as err:
+            self.logger.debug("Error fetching HLS metadata: %s", err)
