+++ music_assistant/controllers/streams/smart_fades/mixer.py
@@ -9,10 +9,7 @@ from music_assistant.controllers.streams.smart_fades.fades import (
     SmartFade,
     StandardCrossFade,
 )
-from music_assistant.helpers.audio import (
-    align_audio_to_frame_boundary,
-    strip_silence,
-)
+from music_assistant.helpers.audio import align_audio_to_frame_boundary, strip_silence
 from music_assistant.models.smart_fades import (
     SmartFadesAnalysis,
     SmartFadesAnalysisFragment,
@@ -23,7 +20,7 @@ if TYPE_CHECKING:
     from music_assistant_models.media_items import AudioFormat
     from music_assistant_models.streamdetails import StreamDetails
 
-    from music_assistant.controllers.streams.streams_controller import StreamsController
+    from music_assistant.controllers.streams.controller import StreamsController
 
 
 class SmartFadesMixer:
@@ -54,7 +51,6 @@ class SmartFadesMixer:
         if mode == SmartFadesMode.STANDARD_CROSSFADE:
             # strip silence from end of audio of fade_out_part
             fade_out_part = await strip_silence(
-                self.streams.mass,
                 fade_out_part,
                 pcm_format=pcm_format,
                 reverse=True,
@@ -63,7 +59,6 @@ class SmartFadesMixer:
             fade_out_part = align_audio_to_frame_boundary(fade_out_part, pcm_format)
             # strip silence from begin of audio of fade_in_part
             fade_in_part = await strip_silence(
-                self.streams.mass,
                 fade_in_part,
                 pcm_format=pcm_format,
                 reverse=False,
