+++ music_assistant/helpers/audio.py
@@ -2,81 +2,44 @@
 
 from __future__ import annotations
 
-import asyncio
 import logging
-import os
 import re
 import struct
-import time
+import urllib.parse
 from collections.abc import AsyncGenerator
-from contextlib import asynccontextmanager
-from functools import partial
 from io import BytesIO
-from typing import TYPE_CHECKING, Any, Final, cast
+from typing import TYPE_CHECKING, Final
 
-import aiofiles
-import aiohttp
-import shortuuid
-from aiohttp import ClientConnectorSSLError, ClientTimeout
-from music_assistant_models.dsp import DSPConfig, DSPDetails, DSPState
 from music_assistant_models.enums import (
     ContentType,
     MediaType,
     PlayerFeature,
     PlayerType,
-    StreamType,
     VolumeNormalizationMode,
 )
-from music_assistant_models.errors import (
-    AudioError,
-    InvalidDataError,
-    MediaNotFoundError,
-    MusicAssistantError,
-    ProviderPermissionDenied,
-    ProviderUnavailableError,
-    RetriesExhausted,
-)
-from music_assistant_models.media_items import AudioFormat
-from music_assistant_models.streamdetails import MultiPartPath, StreamMetadata
+from music_assistant_models.errors import InvalidDataError
+from music_assistant_models.streamdetails import MultiPartPath
 
 from music_assistant.constants import (
-    CONF_ENTRY_OUTPUT_LIMITER,
-    CONF_ENTRY_VOLUME_NORMALIZATION_TARGET,
-    CONF_OUTPUT_CHANNELS,
     CONF_VOLUME_NORMALIZATION,
     CONF_VOLUME_NORMALIZATION_RADIO,
-    CONF_VOLUME_NORMALIZATION_TARGET,
     CONF_VOLUME_NORMALIZATION_TRACKS,
     MASS_LOGGER_NAME,
     VERBOSE_LOG_LEVEL,
 )
 from music_assistant.helpers.json import JSON_DECODE_EXCEPTIONS, json_loads
-from music_assistant.helpers.throttle_retry import BYPASS_THROTTLER
-from music_assistant.helpers.util import clean_stream_title, remove_file
-from music_assistant.providers.sync_group.constants import SGP_PREFIX
 
-from . import ssl as ssl_util
-from .audio_buffer import AudioBuffer
-from .dsp import filter_to_ffmpeg_params
-from .ffmpeg import FFMpeg, get_ffmpeg_args, get_ffmpeg_stream
-from .ogg_handler import get_chained_ogg_stream
-from .playlists import IsHLSPlaylist, PlaylistItem, fetch_playlist, parse_m3u
+from .ffmpeg import get_ffmpeg_args
 from .process import AsyncProcess, communicate
-from .util import detect_charset
 
 if TYPE_CHECKING:
     from music_assistant_models.config_entries import CoreConfig, PlayerConfig
-    from music_assistant_models.queue_item import QueueItem
+    from music_assistant_models.media_items import AudioFormat
     from music_assistant_models.streamdetails import StreamDetails
 
-    from music_assistant.mass import MusicAssistant
-    from music_assistant.models.music_provider import MusicProvider
     from music_assistant.models.player import Player
-    from music_assistant.providers.sync_group import SyncGroupPlayer
 
-LOGGER = logging.getLogger(f"{MASS_LOGGER_NAME}.audio")
-
-# ruff: noqa: PLR0915
+LOGGER = logging.getLogger(f"{MASS_LOGGER_NAME}.helpers.audio")
 
 HTTP_HEADERS = {"User-Agent": "Lavf/60.16.100.MusicAssistant"}
 HTTP_HEADERS_ICY = {**HTTP_HEADERS, "Icy-MetaData": "1"}
@@ -103,6 +66,20 @@ def get_mime_type(format_str: str) -> str:
     return f"audio/{format_str}"
 
 
+def parse_pcm_info(content_type: str) -> tuple[int, int, int]:
+    """Parse PCM info from a codec/content_type string.
+
+    :param content_type: Content type string like "pcm;codec=pcm;rate=44100;bitrate=16;channels=2".
+    """
+    params = (
+        dict(urllib.parse.parse_qsl(content_type.replace(";", "&"))) if ";" in content_type else {}
+    )
+    sample_rate = int(params.get("rate", 44100))
+    sample_size = int(params.get("bitrate", 16))
+    channels = int(params.get("channels", 2))
+    return (sample_rate, sample_size, channels)
+
+
 CACHE_CATEGORY_RESOLVED_RADIO_URL: Final[int] = 100
 CACHE_PROVIDER: Final[str] = "audio"
 
@@ -126,12 +103,16 @@ def align_audio_to_frame_boundary(audio_data: bytes, pcm_format: AudioFormat) ->
 
 
 async def strip_silence(
-    mass: MusicAssistant,  # noqa: ARG001
     audio_data: bytes,
     pcm_format: AudioFormat,
     reverse: bool = False,
 ) -> bytes:
-    """Strip silence from begin or end of pcm audio using ffmpeg."""
+    """Strip silence from begin or end of pcm audio using ffmpeg.
+
+    :param audio_data: Raw PCM audio data.
+    :param pcm_format: AudioFormat of the audio data.
+    :param reverse: If True, strip from end instead of beginning.
+    """
     args = ["ffmpeg", "-hide_banner", "-loglevel", "quiet"]
     args += [
         "-acodec",
@@ -145,22 +126,20 @@ async def strip_silence(
         "-i",
         "-",
     ]
-    # filter args
     if reverse:
         args += [
             "-af",
-            "areverse,atrim=start=0.2,silenceremove=start_periods=1:start_silence=0.1:start_threshold=0.02,areverse",
+            "areverse,atrim=start=0.2,silenceremove=start_periods=1"
+            ":start_silence=0.1:start_threshold=0.02,areverse",
         ]
     else:
         args += [
             "-af",
             "atrim=start=0.2,silenceremove=start_periods=1:start_silence=0.1:start_threshold=0.02",
         ]
-    # output args
     args += ["-f", pcm_format.content_type.value, "-"]
     _returncode, stripped_data, _stderr = await communicate(args, audio_data)
 
-    # return stripped audio
     bytes_stripped = len(audio_data) - len(stripped_data)
     if LOGGER.isEnabledFor(VERBOSE_LOG_LEVEL):
         seconds_stripped = round(bytes_stripped / pcm_format.pcm_sample_size, 2)
@@ -175,595 +154,11 @@ async def strip_silence(
     return stripped_data
 
 
-def get_player_dsp_details(
-    mass: MusicAssistant, player: Player, group_preventing_dsp: bool = False
-) -> DSPDetails:
-    """Return DSP details of single a player.
-
-    This will however not check if the queried player is part of a group.
-    The caller is responsible for passing the result of is_grouping_preventing_dsp of
-    the leader/PlayerGroup as the group_preventing_dsp argument in such cases.
-    """
-    dsp_config = mass.config.get_player_dsp_config(player.player_id)
-    dsp_state = DSPState.ENABLED if dsp_config.enabled else DSPState.DISABLED
-    if dsp_state == DSPState.ENABLED and (
-        group_preventing_dsp or is_grouping_preventing_dsp(player)
-    ):
-        dsp_state = DSPState.DISABLED_BY_UNSUPPORTED_GROUP
-        dsp_config = DSPConfig(enabled=False)
-    elif dsp_state == DSPState.DISABLED:
-        # DSP is disabled by the user, remove all filters
-        dsp_config = DSPConfig(enabled=False)
-
-    # remove disabled filters
-    dsp_config.filters = [x for x in dsp_config.filters if x.enabled]
-
-    output_limiter = is_output_limiter_enabled(mass, player)
-    return DSPDetails(
-        state=dsp_state,
-        input_gain=dsp_config.input_gain,
-        filters=dsp_config.filters,
-        output_gain=dsp_config.output_gain,
-        output_limiter=output_limiter,
-        output_format=player.extra_data.get("output_format", None),
-    )
-
-
-def get_stream_dsp_details(
-    mass: MusicAssistant,
-    queue_id: str,
-) -> dict[str, DSPDetails]:
-    """Return DSP details of all players playing this queue, keyed by player_id."""
-    player = mass.players.get_player(queue_id)
-    dsp: dict[str, DSPDetails] = {}
-    assert player is not None  # for type checking
-    group_preventing_dsp = is_grouping_preventing_dsp(player)
-    output_format = None
-    is_external_group = False
-
-    if player.player_id.startswith(SGP_PREFIX):
-        if group_preventing_dsp:
-            sgp_player = cast("SyncGroupPlayer", player)
-            if sync_leader := sgp_player.sync_leader:
-                output_format = sync_leader.extra_data.get("output_format", None)
-    else:
-        # We only add real players (so skip the PlayerGroups as they only sync containing players)
-        details = get_player_dsp_details(mass, player)
-        dsp[player.player_id] = details
-        if group_preventing_dsp:
-            # The leader is responsible for sending the (combined) audio stream, so get
-            # the output format from the leader.
-            output_format = player.extra_data.get("output_format", None)
-        is_external_group = player.state.type in (PlayerType.GROUP, PlayerType.STEREO_PAIR)
-
-    # We don't enumerate all group members in case this group is externally created
-    # (e.g. a Chromecast group from the Google Home app)
-    if player and player.state.group_members and not is_external_group:
-        # grouped playback, get DSP details for each player in the group
-        for child_id in player.state.group_members:
-            # skip if we already have the details (so if it's the group leader)
-            if child_id in dsp:
-                continue
-            if child_player := mass.players.get_player(child_id):
-                dsp[child_id] = get_player_dsp_details(
-                    mass, child_player, group_preventing_dsp=group_preventing_dsp
-                )
-                if group_preventing_dsp:
-                    # Use the correct format from the group leader, since
-                    # this player is part of a group that does not support
-                    # multi device DSP processing.
-                    dsp[child_id].output_format = output_format
-    return dsp
-
-
-async def get_stream_details(
-    mass: MusicAssistant,
-    queue_item: QueueItem,
-    seek_position: int = 0,
-    fade_in: bool = False,
-    prefer_album_loudness: bool = False,
-) -> StreamDetails:
-    """
-    Get streamdetails for the given QueueItem.
-
-    This is called just-in-time when a PlayerQueue wants a MediaItem to be played.
-    Do not try to request streamdetails too much in advance as this is expiring data.
-    """
-    streamdetails: StreamDetails | None = None
-    time_start = time.time()
-    LOGGER.debug("Getting streamdetails for %s", queue_item.uri)
-    if seek_position and (queue_item.media_type == MediaType.RADIO or not queue_item.duration):
-        LOGGER.warning("seeking is not possible on duration-less streams!")
-        seek_position = 0
-
-    if not queue_item.media_item and not queue_item.streamdetails:
-        # in case of a non-media item queue item, the streamdetails should already be provided
-        # this should not happen, but guard it just in case
-        raise MediaNotFoundError(
-            f"Unable to retrieve streamdetails for {queue_item.name} ({queue_item.uri})"
-        )
-    buffer: AudioBuffer | None = None
-    if queue_item.streamdetails and (
-        (queue_item.streamdetails.created_at + queue_item.streamdetails.expiration) > time.time()
-        or ((buffer := queue_item.streamdetails.buffer) and buffer.is_valid(seek_position))
-    ):
-        # already got a fresh/unused (or unexpired) streamdetails
-        streamdetails = queue_item.streamdetails
-    else:
-        # need to (re)create streamdetails
-        # retrieve streamdetails from provider
-
-        media_item = queue_item.media_item
-        assert media_item is not None  # for type checking
-        preferred_providers: list[str] = []
-        if (
-            (queue := mass.player_queues.get(queue_item.queue_id))
-            and queue.userid
-            and (playback_user := await mass.webserver.auth.get_user(queue.userid))
-            and playback_user.provider_filter
-        ):
-            # handle steering into user preferred providerinstance
-            preferred_providers = playback_user.provider_filter
-        else:
-            preferred_providers = [x.provider_instance for x in media_item.provider_mappings]
-        for allow_other_provider in (False, True):
-            if streamdetails:
-                break
-            # sort by quality and check item's availability
-            for prov_media in sorted(
-                media_item.provider_mappings, key=lambda x: x.quality or 0, reverse=True
-            ):
-                if not prov_media.available:
-                    LOGGER.debug(f"Skipping unavailable {prov_media}")
-                    continue
-                if (
-                    not allow_other_provider
-                    and prov_media.provider_instance not in preferred_providers
-                ):
-                    continue
-                # guard that provider is available
-                music_prov = mass.get_provider(prov_media.provider_instance)
-                if TYPE_CHECKING:  # avoid circular import
-                    assert isinstance(music_prov, MusicProvider)
-                if not music_prov:
-                    LOGGER.debug(f"Skipping {prov_media} - provider not available")
-                    continue  # provider not available ?
-                # get streamdetails from provider
-                try:
-                    BYPASS_THROTTLER.set(True)
-                    streamdetails = await music_prov.get_stream_details(
-                        prov_media.item_id, media_item.media_type
-                    )
-                except MusicAssistantError as err:
-                    LOGGER.warning(str(err))
-                else:
-                    break
-                finally:
-                    BYPASS_THROTTLER.set(False)
-
-        if not streamdetails:
-            msg = f"Unable to retrieve streamdetails for {queue_item.name} ({queue_item.uri})"
-            raise MediaNotFoundError(msg)
-
-        # work out how to handle radio stream
-        if (
-            streamdetails.stream_type in (StreamType.ICY, StreamType.HLS, StreamType.HTTP)
-            and streamdetails.media_type == MediaType.RADIO
-            and isinstance(streamdetails.path, str)
-        ):
-            resolved_url, stream_type = await resolve_radio_stream(mass, streamdetails.path)
-            streamdetails.path = resolved_url
-            streamdetails.stream_type = stream_type
-            # Set up metadata monitoring callback for HLS radio streams, if not already set
-            if stream_type == StreamType.HLS and not streamdetails.stream_metadata_update_callback:
-                streamdetails.stream_metadata_update_callback = partial(
-                    _update_hls_radio_metadata, mass
-                )
-                streamdetails.stream_metadata_update_interval = 5
-        # handle volume normalization details
-        if result := await mass.music.get_loudness(
-            streamdetails.item_id,
-            streamdetails.provider,
-            media_type=queue_item.media_type,
-        ):
-            streamdetails.loudness = result[0]
-            streamdetails.loudness_album = result[1]
-
-    # set queue_id on the streamdetails so we know what is being streamed
-    streamdetails.queue_id = queue_item.queue_id
-    # handle skip/fade_in details
-    streamdetails.seek_position = seek_position
-    streamdetails.fade_in = fade_in
-    if not streamdetails.duration:
-        streamdetails.duration = queue_item.duration
-    streamdetails.prefer_album_loudness = prefer_album_loudness
-    player_settings = await mass.config.get_player_config(streamdetails.queue_id)
-    core_config = await mass.config.get_core_config("streams")
-    conf_volume_normalization_target = float(
-        str(player_settings.get_value(CONF_VOLUME_NORMALIZATION_TARGET, -17))
-    )
-    # guard against invalid volume normalization values
-    # range and default_value are guaranteed to be set for this constant
-    volume_range = CONF_ENTRY_VOLUME_NORMALIZATION_TARGET.range
-    assert volume_range is not None
-    if (
-        conf_volume_normalization_target < volume_range[0]
-        or conf_volume_normalization_target >= volume_range[1]
-    ):
-        default_val = CONF_ENTRY_VOLUME_NORMALIZATION_TARGET.default_value
-        assert isinstance(default_val, (int, float))
-        conf_volume_normalization_target = float(default_val)
-        LOGGER.warning(
-            "Invalid volume normalization target configured for player %s, "
-            "resetting to default of %s dB",
-            streamdetails.queue_id,
-            CONF_ENTRY_VOLUME_NORMALIZATION_TARGET.default_value,
-        )
-    streamdetails.target_loudness = conf_volume_normalization_target
-    streamdetails.volume_normalization_mode = _get_normalization_mode(
-        core_config, player_settings, streamdetails
-    )
-
-    # attach the DSP details of all group members
-    streamdetails.dsp = get_stream_dsp_details(mass, streamdetails.queue_id)
-
-    LOGGER.debug(
-        "retrieved streamdetails for %s in %s milliseconds",
-        queue_item.uri,
-        int((time.time() - time_start) * 1000),
-    )
-    return streamdetails
-
-
-async def get_buffered_media_stream(
-    mass: MusicAssistant,
-    streamdetails: StreamDetails,
-    pcm_format: AudioFormat,
-    seek_position: int = 0,
-    filter_params: list[str] | None = None,
-) -> AsyncGenerator[bytes, None]:
-    """Get audio stream for given media details as raw PCM with buffering."""
-    LOGGER.log(
-        VERBOSE_LOG_LEVEL,
-        "buffered_media_stream: Starting for %s (seek: %s)",
-        streamdetails.uri,
-        seek_position,
-    )
-
-    # checksum based on filter_params
-    checksum = f"{filter_params}"
-
-    async def fill_buffer_task() -> None:
-        """Background task to fill the audio buffer."""
-        chunk_count = 0
-        status = "running"
-        try:
-            async for chunk in get_media_stream(
-                mass, streamdetails, pcm_format, seek_position=0, filter_params=filter_params
-            ):
-                chunk_count += 1
-                await audio_buffer.put(chunk)
-                # Yield to event loop to prevent blocking warnings
-                await asyncio.sleep(0)
-            # Only set EOF if we completed successfully
-            await audio_buffer.set_eof()
-        except asyncio.CancelledError:
-            status = "cancelled"
-            raise
-        except Exception:
-            status = "aborted with error"
-            raise
-        finally:
-            LOGGER.log(
-                VERBOSE_LOG_LEVEL,
-                "fill_buffer_task: %s (%s chunks) for %s",
-                status,
-                chunk_count,
-                streamdetails.uri,
-            )
-
-    # check for existing buffer and reuse if possible
-    existing_buffer: AudioBuffer | None = streamdetails.buffer
-    if existing_buffer is not None:
-        if not existing_buffer.is_valid(checksum, seek_position):
-            LOGGER.log(
-                VERBOSE_LOG_LEVEL,
-                "buffered_media_stream: Existing buffer invalid for %s (seek: %s, discarded: %s)",
-                streamdetails.uri,
-                seek_position,
-                existing_buffer._discarded_chunks,
-            )
-            buffer_to_clear = existing_buffer
-            streamdetails.buffer = None
-            existing_buffer = None
-            await asyncio.shield(buffer_to_clear.clear())
-        else:
-            LOGGER.debug(
-                "buffered_media_stream: Reusing existing buffer for %s - "
-                "available: %ss, seek: %s, discarded: %s",
-                streamdetails.uri,
-                existing_buffer.seconds_available,
-                seek_position,
-                existing_buffer._discarded_chunks,
-            )
-            audio_buffer = existing_buffer
-
-    if not existing_buffer and seek_position > 60:
-        # If seeking into the track and no valid buffer exists,
-        # just start a normal stream without buffering,
-        # otherwise we would need to fill the buffer up to the seek position first
-        # which is not efficient.
-        LOGGER.debug(
-            "buffered_media_stream: No existing buffer and seek >60s for %s, "
-            "starting normal (unbuffered) stream",
-            streamdetails.uri,
-        )
-        media_stream = get_media_stream(
-            mass,
-            streamdetails,
-            pcm_format,
-            seek_position=seek_position,
-            filter_params=filter_params,
-        )
-        completed = False
-        try:
-            async for chunk in media_stream:
-                yield chunk
-            completed = True
-        finally:
-            if not completed:
-                await asyncio.shield(media_stream.aclose())
-        return
-
-    if not existing_buffer:
-        # create new audio buffer and start fill task
-        LOGGER.debug(
-            "buffered_media_stream: Creating new buffer for %s",
-            streamdetails.uri,
-        )
-        audio_buffer = AudioBuffer(pcm_format, checksum)
-        streamdetails.buffer = audio_buffer
-        task = mass.loop.create_task(fill_buffer_task())
-        audio_buffer.attach_producer_task(task)
-
-    # special case: pcm format mismatch, resample on the fly
-    # this may happen in some special situations such as crossfading
-    # and its a bit of a waste to throw away the existing buffer
-    if audio_buffer.pcm_format != pcm_format:
-        LOGGER.info(
-            "buffered_media_stream: pcm format mismatch, resampling on the fly for %s - "
-            "buffer format: %s - requested format: %s",
-            streamdetails.uri,
-            audio_buffer.pcm_format,
-            pcm_format,
-        )
-        async for chunk in get_ffmpeg_stream(
-            audio_input=audio_buffer.iter(seek_position=seek_position),
-            input_format=audio_buffer.pcm_format,
-            output_format=pcm_format,
-        ):
-            yield chunk
-        return
-
-    # yield data from the buffer
-    chunk_count = 0
-    try:
-        async for chunk in audio_buffer.iter(seek_position=seek_position):
-            chunk_count += 1
-            yield chunk
-    finally:
-        LOGGER.log(
-            VERBOSE_LOG_LEVEL,
-            "buffered_media_stream: Completed, yielded %s chunks",
-            chunk_count,
-        )
-
-
-async def get_media_stream(
-    mass: MusicAssistant,
-    streamdetails: StreamDetails,
-    pcm_format: AudioFormat,
-    seek_position: int = 0,
-    filter_params: list[str] | None = None,
-) -> AsyncGenerator[bytes, None]:
-    """Get audio stream for given media details as raw PCM."""
-    logger = LOGGER.getChild("media_stream")
-    logger.log(VERBOSE_LOG_LEVEL, "Starting media stream for %s", streamdetails.uri)
-    extra_input_args = streamdetails.extra_input_args or []
-
-    # work out audio source for these streamdetails
-    audio_source: str | AsyncGenerator[bytes, None]
-    stream_type = streamdetails.stream_type
-    if stream_type == StreamType.CUSTOM:
-        music_prov = mass.get_provider(streamdetails.provider)
-        if TYPE_CHECKING:  # avoid circular import
-            assert isinstance(music_prov, MusicProvider)
-        audio_source = music_prov.get_audio_stream(
-            streamdetails,
-            seek_position=seek_position if streamdetails.can_seek else 0,
-        )
-        seek_position = 0 if streamdetails.can_seek else seek_position
-    elif stream_type == StreamType.ICY:
-        assert isinstance(streamdetails.path, str)  # for type checking
-        audio_source = get_icy_radio_stream(mass, streamdetails.path, streamdetails)
-        seek_position = 0  # seeking not possible on radio streams
-    elif stream_type == StreamType.IN_BAND:
-        assert isinstance(streamdetails.path, str)  # for type checking
-
-        # For IN_BAND (OGG/Opus) radio streams, use chained OGG handler.
-        # This handles the chained OGG format by stitching logical bitstreams together
-        # so FFmpeg sees a single continuous stream. Metadata is extracted in-band.
-        def _on_inband_metadata(metadata: dict[str, str]) -> None:
-            """Handle metadata extracted from the OGG stream."""
-            title = metadata.get("title", "")
-            artist = metadata.get("artist", "")
-            album = metadata.get("album", "")
-            if not artist and " - " in title:
-                artist, title = title.split(" - ", 1)
-            if title or artist:
-                if artist and title:
-                    stream_title = f"{artist} - {title}"
-                elif title:
-                    stream_title = title
-                else:
-                    stream_title = artist
-                cleaned_title = clean_stream_title(stream_title)
-                if cleaned_title and cleaned_title != streamdetails.stream_title:
-                    LOGGER.log(
-                        VERBOSE_LOG_LEVEL,
-                        "In-band metadata: %s",
-                        cleaned_title,
-                    )
-                    streamdetails.stream_title = cleaned_title
-                    streamdetails.stream_metadata = StreamMetadata(
-                        title=title or cleaned_title,
-                        artist=artist or None,
-                        album=album or None,
-                    )
-
-        audio_source = get_chained_ogg_stream(
-            mass, streamdetails.path, metadata_callback=_on_inband_metadata
-        )
-        seek_position = 0  # seeking not possible on radio streams
-    elif stream_type == StreamType.HLS:
-        assert isinstance(streamdetails.path, str)  # for type checking
-        substream = await get_hls_substream(mass, streamdetails.path)
-        audio_source = substream.path
-        if streamdetails.media_type == MediaType.RADIO:
-            # HLS streams (especially the BBC) struggle when they're played directly
-            # with ffmpeg, where they just stop after some minutes,
-            # so we tell ffmpeg to loop around in this case.
-            extra_input_args += ["-stream_loop", "-1", "-re"]
-    else:
-        # all other stream types (HTTP, FILE, etc)
-        if stream_type == StreamType.ENCRYPTED_HTTP:
-            assert streamdetails.decryption_key is not None  # for type checking
-            extra_input_args += ["-decryption_key", streamdetails.decryption_key]
-        if isinstance(streamdetails.path, list):
-            # multi part stream
-            audio_source = get_multi_file_stream(mass, streamdetails, seek_position)
-            seek_position = 0  # handled by get_multi_file_stream
-        else:
-            # regular single file/url stream
-            assert isinstance(streamdetails.path, str)  # for type checking
-            audio_source = streamdetails.path
-
-    # handle seek support
-    if seek_position and streamdetails.duration and streamdetails.allow_seek:
-        extra_input_args += ["-ss", str(int(seek_position))]
-
-    bytes_sent = 0
-    finished = False
-    cancelled = False
-    first_chunk_received = False
-    ffmpeg_loglevel = "debug" if LOGGER.isEnabledFor(VERBOSE_LOG_LEVEL) else "info"
-    ffmpeg_proc = FFMpeg(
-        audio_input=audio_source,
-        input_format=streamdetails.audio_format,
-        output_format=pcm_format,
-        filter_params=filter_params,
-        extra_input_args=extra_input_args,
-        collect_log_history=True,
-        loglevel=ffmpeg_loglevel,
-    )
-
-    try:
-        await ffmpeg_proc.start()
-        assert ffmpeg_proc.proc is not None  # for type checking
-        logger.debug(
-            "Started media stream for %s - using streamtype: %s - pcm format: %s - ffmpeg PID: %s",
-            streamdetails.uri,
-            streamdetails.stream_type,
-            pcm_format.content_type.value,
-            ffmpeg_proc.proc.pid,
-        )
-        stream_start = mass.loop.time()
-
-        chunk_size = get_chunksize(pcm_format, 1)
-        async for chunk in ffmpeg_proc.iter_chunked(chunk_size):
-            if not first_chunk_received:
-                # At this point ffmpeg has started and should now know the codec used
-                # for encoding the audio.
-                first_chunk_received = True
-                streamdetails.audio_format.codec_type = ffmpeg_proc.input_format.codec_type
-                logger.debug(
-                    "First chunk received after %.2f seconds (codec detected: %s)",
-                    mass.loop.time() - stream_start,
-                    ffmpeg_proc.input_format.codec_type,
-                )
-            yield chunk
-            bytes_sent += len(chunk)
-
-        # end of audio/track reached
-        logger.debug("End of ffmpeg output stream reached for %s", streamdetails.uri)
-        # wait until stderr also completed reading
-        await ffmpeg_proc.wait_with_timeout(5)
-        logger.debug(
-            "FFmpeg process ended with return code %s for %s",
-            ffmpeg_proc.returncode,
-            streamdetails.uri,
-        )
-        if ffmpeg_proc.returncode not in (0, None):
-            log_trail = "\n".join(list(ffmpeg_proc.log_history)[-5:])
-            raise AudioError(f"FFMpeg exited with code {ffmpeg_proc.returncode}: {log_trail}")
-        if bytes_sent == 0:
-            # edge case: no audio data was received at all
-            raise AudioError("No audio was received")
-        finished = True
-    except (Exception, GeneratorExit, asyncio.CancelledError) as err:
-        if isinstance(err, asyncio.CancelledError | GeneratorExit):
-            # we were cancelled, just raise
-            cancelled = True
-            raise
-        # dump the last 10 lines of the log in case of an unclean exit
-        logger.warning("\n".join(list(ffmpeg_proc.log_history)[-10:]))
-        raise AudioError(f"Error while streaming: {err}") from err
-    finally:
-        # always ensure close is called which also handles all cleanup
-        await ffmpeg_proc.close()
-        # determine how many seconds we've received
-        # for pcm output we can calculate this easily
-        seconds_received = bytes_sent / pcm_format.pcm_sample_size if bytes_sent else 0
-        # store accurate duration
-        if finished and not seek_position and seconds_received:
-            streamdetails.duration = int(seconds_received)
-
-        logger.log(
-            VERBOSE_LOG_LEVEL,
-            "stream %s (with code %s) for %s",
-            "cancelled" if cancelled else "finished" if finished else "aborted",
-            ffmpeg_proc.returncode,
-            streamdetails.uri,
-        )
-
-        # parse loudnorm data if we have that collected (and enabled)
-        if (
-            (streamdetails.loudness is None or finished)
-            and streamdetails.volume_normalization_mode == VolumeNormalizationMode.DYNAMIC
-            and (finished or (seconds_received >= 300))
-        ):
-            # if dynamic volume normalization is enabled
-            # the loudnorm filter will output the measurement in the log,
-            # so we can use that directly instead of analyzing the audio
-            logger.log(VERBOSE_LOG_LEVEL, "Collecting loudness measurement...")
-            if loudness_details := parse_loudnorm(" ".join(ffmpeg_proc.log_history)):
-                logger.debug(
-                    "Loudness measurement for %s: %s dB",
-                    streamdetails.uri,
-                    loudness_details,
-                )
-                mass.create_task(
-                    mass.music.set_loudness(
-                        streamdetails.item_id,
-                        streamdetails.provider,
-                        loudness_details,
-                        media_type=streamdetails.media_type,
-                    )
-                )
-
-
 def create_wave_header(
-    samplerate: int = 44100, channels: int = 2, bitspersample: int = 16, duration: int | None = None
+    samplerate: int = 44100,
+    channels: int = 2,
+    bitspersample: int = 16,
+    duration: int | None = None,
 ) -> bytes:
     """Generate a wave header from given params."""
     file = BytesIO()
@@ -817,216 +212,6 @@ def create_wave_header(
     return file.getvalue()
 
 
-@asynccontextmanager
-async def _connect_radio_stream(
-    mass: MusicAssistant, url: str, **kwargs: Any
-) -> AsyncGenerator[Any, None]:
-    """Connect to a radio stream URL with fallback for legacy SSL/TLS configurations.
-
-    Some radio servers use outdated TLS configurations that reject modern cipher suites.
-    Since radio streams are public broadcast content, relaxing cipher requirements is acceptable.
-
-    :param mass: The MusicAssistant instance.
-    :param url: The radio stream URL to connect to.
-    :param kwargs: Additional keyword arguments passed to aiohttp get().
-    """
-    try:
-        async with mass.http_session_no_ssl.get(url, **kwargs) as resp:
-            yield resp
-    except ClientConnectorSSLError:
-        LOGGER.info(
-            "SSL handshake failed for %s, retrying with permissive cipher configuration",
-            url,
-        )
-        insecure_ssl_context = ssl_util.client_context_no_verify(ssl_util.SSLCipherList.INSECURE)
-        async with mass.http_session_no_ssl.get(url, ssl=insecure_ssl_context, **kwargs) as resp:
-            yield resp
-
-
-async def resolve_radio_stream(mass: MusicAssistant, url: str) -> tuple[str, StreamType]:
-    """
-    Resolve a streaming radio URL.
-
-    Unwraps any playlists if needed.
-    Determines if the stream supports ICY metadata or in-band metadata.
-
-    Returns tuple;
-    - unfolded URL as string
-    - StreamType to determine ICY (radio), HLS, or IN_BAND stream.
-    """
-    if cache := await mass.cache.get(
-        key=url, provider=CACHE_PROVIDER, category=CACHE_CATEGORY_RESOLVED_RADIO_URL
-    ):
-        if TYPE_CHECKING:  # for type checking
-            cache = cast("tuple[str, str]", cache)
-        return (cache[0], StreamType(cache[1]))
-    stream_type = StreamType.HTTP
-    resolved_url = url
-    timeout = ClientTimeout(total=None, connect=10, sock_read=5)
-    try:
-        async with _connect_radio_stream(
-            mass, url, headers=HTTP_HEADERS_ICY, allow_redirects=True, timeout=timeout
-        ) as resp:
-            headers = resp.headers
-            resp.raise_for_status()
-            if not resp.headers:
-                raise InvalidDataError("no headers found")
-        content_type = headers.get("content-type", "")
-        if headers.get("icy-metaint") is not None:
-            stream_type = StreamType.ICY
-        elif content_type in ("application/ogg", "audio/ogg"):
-            # Ogg streams (Opus/Vorbis) have in-band metadata via Vorbis comments
-            stream_type = StreamType.IN_BAND
-        if (
-            url.endswith((".m3u", ".m3u8", ".pls"))
-            or ".m3u?" in url
-            or ".m3u8?" in url
-            or ".pls?" in url
-            or "audio/x-mpegurl" in headers.get("content-type", "")
-            or "audio/x-scpls" in headers.get("content-type", "")
-        ):
-            # url is playlist, we need to unfold it
-            try:
-                substreams = await fetch_playlist(mass, url)
-                if not any(x for x in substreams if x.length):
-                    for line in substreams:
-                        if not line.is_url:
-                            continue
-                        # unfold first url of playlist
-                        return await resolve_radio_stream(mass, line.path)
-                    raise InvalidDataError("No content found in playlist")
-            except IsHLSPlaylist:
-                stream_type = StreamType.HLS
-
-    except Exception as err:
-        LOGGER.warning("Error while parsing radio URL %s: %s", url, str(err))
-        return (url, stream_type)
-
-    result = (resolved_url, stream_type)
-    cache_expiration = 3600 * 3
-    await mass.cache.set(
-        url,
-        result,
-        expiration=cache_expiration,
-        provider=CACHE_PROVIDER,
-        category=CACHE_CATEGORY_RESOLVED_RADIO_URL,
-    )
-    return result
-
-
-async def get_icy_radio_stream(
-    mass: MusicAssistant, url: str, streamdetails: StreamDetails
-) -> AsyncGenerator[bytes, None]:
-    """Get (radio) audio stream from HTTP, including ICY metadata retrieval."""
-    timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
-    LOGGER.debug("Start streaming radio with ICY metadata from url %s", url)
-    async with _connect_radio_stream(
-        mass, url, allow_redirects=True, headers=HTTP_HEADERS_ICY, timeout=timeout
-    ) as resp:
-        headers = resp.headers
-        meta_int = int(headers["icy-metaint"])
-        while True:
-            try:
-                yield await resp.content.readexactly(meta_int)
-                meta_byte = await resp.content.readexactly(1)
-                if meta_byte == b"\x00":
-                    continue
-                meta_length = ord(meta_byte) * 16
-                meta_data = await resp.content.readexactly(meta_length)
-            except asyncio.exceptions.IncompleteReadError:
-                break
-            if not meta_data:
-                continue
-            meta_data = meta_data.rstrip(b"\0")
-            stream_title_re = re.search(rb"StreamTitle='([^']*)';", meta_data)
-            if not stream_title_re:
-                continue
-            try:
-                # in 99% of the cases the stream title is utf-8 encoded
-                stream_title = stream_title_re.group(1).decode("utf-8")
-            except UnicodeDecodeError:
-                # fallback to iso-8859-1
-                stream_title = stream_title_re.group(1).decode("iso-8859-1", errors="replace")
-            cleaned_stream_title = clean_stream_title(stream_title)
-            if cleaned_stream_title != streamdetails.stream_title:
-                LOGGER.log(
-                    VERBOSE_LOG_LEVEL,
-                    "ICY Radio streamtitle original: %s",
-                    stream_title,
-                )
-                LOGGER.log(
-                    VERBOSE_LOG_LEVEL,
-                    "ICY Radio streamtitle cleaned: %s",
-                    cleaned_stream_title,
-                )
-                streamdetails.stream_title = cleaned_stream_title
-
-
-async def get_reconnecting_radio_stream(
-    mass: MusicAssistant, url: str
-) -> AsyncGenerator[bytes, None]:
-    """Yield continuous radio stream data, automatically reconnecting on disconnect.
-
-    :param mass: MusicAssistant instance.
-    :param url: URL of the radio stream.
-    """
-    timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
-    reconnect_count = 0
-    max_reconnects = 1000  # Allow many reconnects for long-running radio
-
-    while reconnect_count <= max_reconnects:
-        try:
-            async with _connect_radio_stream(
-                mass, url, allow_redirects=True, headers=HTTP_HEADERS, timeout=timeout
-            ) as resp:
-                chunk_count = 0
-                async for chunk in resp.content.iter_any():
-                    chunk_count += 1
-                    yield chunk
-
-                # Connection closed normally - reconnect
-                LOGGER.debug(
-                    "Radio stream connection closed after %d chunks, reconnecting... "
-                    "(reconnect #%d)",
-                    chunk_count,
-                    reconnect_count,
-                )
-                reconnect_count += 1
-                await asyncio.sleep(0.1)  # Brief delay before reconnect
-
-        except asyncio.CancelledError:
-            LOGGER.debug("Radio stream cancelled for %s", url)
-            raise
-        except (
-            aiohttp.ClientConnectionError,
-            aiohttp.ClientPayloadError,
-            aiohttp.ServerDisconnectedError,
-        ) as err:
-            # Transient network errors - retry
-            LOGGER.warning(
-                "Radio stream error (reconnect #%d): %s",
-                reconnect_count,
-                err,
-            )
-            reconnect_count += 1
-            if reconnect_count > max_reconnects:
-                raise RetriesExhausted(
-                    f"Radio stream failed after {max_reconnects} reconnects: {err}"
-                ) from err
-            await asyncio.sleep(0.5)
-        except aiohttp.ClientResponseError as err:
-            if err.status == 404:
-                raise MediaNotFoundError(f"Radio stream not found: {url}") from err
-            if err.status == 403:
-                raise ProviderPermissionDenied(f"Radio stream access denied: {url}") from err
-            # Other HTTP errors (5xx etc) - could be temporary
-            raise ProviderUnavailableError(
-                f"Radio stream returned HTTP {err.status}: {err}"
-            ) from err
-
-    LOGGER.warning("Radio stream reached max reconnects (%d) for %s", max_reconnects, url)
-
-
 def parse_extinf_metadata(extinf_line: str) -> dict[str, str]:
     """
     Parse metadata from HLS EXTINF line.
@@ -1049,249 +234,9 @@ def parse_extinf_metadata(extinf_line: str) -> dict[str, str]:
     return metadata
 
 
-async def _update_hls_radio_metadata(
-    mass: MusicAssistant,
-    streamdetails: StreamDetails,
-    elapsed_time: int,  # noqa: ARG001
-) -> None:
-    """
-    Update HLS radio stream metadata by fetching the playlist.
-
-    Fetches the HLS playlist and extracts metadata from EXTINF lines.
-
-    :param mass: MusicAssistant instance
-    :param streamdetails: StreamDetails object to update with metadata
-    :param elapsed_time: Current playback position in seconds (unused for live radio)
-    """
-    try:
-        # Get the actual media playlist URL from cache or resolve it
-        # We cache the media_playlist_url in streamdetails.data to avoid re-resolving
-        if streamdetails.data is None:
-            streamdetails.data = {}
-        media_playlist_url = streamdetails.data.get("hls_media_playlist_url")
-        if not media_playlist_url:
-            try:
-                assert isinstance(streamdetails.path, str)  # for type checking
-                substream = await get_hls_substream(mass, streamdetails.path)
-                media_playlist_url = substream.path
-                streamdetails.data["hls_media_playlist_url"] = media_playlist_url
-            except Exception as err:
-                LOGGER.warning(
-                    "Failed to resolve HLS substream for metadata monitoring: %s",
-                    err,
-                )
-                return
-
-        # Fetch the media playlist
-        timeout = ClientTimeout(total=0, connect=10, sock_read=30)
-        async with mass.http_session_no_ssl.get(media_playlist_url, timeout=timeout) as resp:
-            resp.raise_for_status()
-            playlist_content = await resp.text()
-
-        # Parse the playlist and look for EXTINF metadata
-        # The most recent segment usually has the current metadata
-        lines = playlist_content.strip().split("\n")
-        for line in reversed(lines):
-            if line.startswith("#EXTINF:"):
-                # Extract metadata from EXTINF line
-                metadata = parse_extinf_metadata(line)
-
-                # Build stream title from title and artist
-                title = metadata.get("title", "")
-                artist = metadata.get("artist", "")
-
-                if title or artist:
-                    # Format as "Artist - Title"
-                    if artist and title:
-                        stream_title = f"{artist} - {title}"
-                    elif title:
-                        stream_title = title
-                    else:
-                        stream_title = artist
-
-                    # Clean the stream title
-                    cleaned_title = clean_stream_title(stream_title)
-
-                    # Only update if changed
-                    if cleaned_title != streamdetails.stream_title and cleaned_title:
-                        LOGGER.log(
-                            VERBOSE_LOG_LEVEL,
-                            "HLS Radio metadata updated: %s",
-                            cleaned_title,
-                        )
-                        streamdetails.stream_title = cleaned_title
-
-                # Only check the most recent EXTINF
-                break
-
-    except Exception as err:
-        LOGGER.debug(
-            "Error fetching HLS metadata: %s",
-            err,
-        )
-
-
-async def get_hls_substream(
-    mass: MusicAssistant,
-    url: str,
-) -> PlaylistItem:
-    """Select the (highest quality) HLS substream for given HLS playlist/URL."""
-    timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
-    # fetch master playlist and select (best) child playlist
-    # https://datatracker.ietf.org/doc/html/draft-pantos-http-live-streaming-19#section-10
-    async with mass.http_session_no_ssl.get(
-        url, allow_redirects=True, headers=HTTP_HEADERS, timeout=timeout
-    ) as resp:
-        resp.raise_for_status()
-        raw_data = await resp.read()
-        encoding = resp.charset or await detect_charset(raw_data)
-        master_m3u_data = raw_data.decode(encoding)
-    substreams = parse_m3u(master_m3u_data)
-    # There is a chance that we did not get a master playlist with subplaylists
-    # but just a single master/sub playlist with the actual audio stream(s)
-    # so we need to detect if the playlist child's contain audio streams or
-    # sub-playlists.
-    if any(
-        x
-        for x in substreams
-        if (x.length or x.path.endswith((".mp4", ".aac")))
-        and not x.path.endswith((".m3u", ".m3u8"))
-    ):
-        return PlaylistItem(path=url, key=substreams[0].key)
-    # sort substreams on best quality (highest bandwidth) when available
-    if any(x for x in substreams if x.stream_info):
-        substreams.sort(
-            key=lambda x: int(
-                x.stream_info.get("BANDWIDTH", "0") if x.stream_info is not None else 0
-            ),
-            reverse=True,
-        )
-    substream = substreams[0]
-    if not substream.path.startswith("http"):
-        # path is relative, stitch it together
-        base_path = url.rsplit("/", 1)[0]
-        substream.path = base_path + "/" + substream.path
-    return substream
-
-
-async def get_http_stream(
-    mass: MusicAssistant,
-    url: str,
-    streamdetails: StreamDetails,
-    seek_position: int = 0,
-    verify_ssl: bool = True,
-) -> AsyncGenerator[bytes, None]:
-    """Get audio stream from HTTP."""
-    LOGGER.debug("Start HTTP stream for %s (seek_position %s)", streamdetails.uri, seek_position)
-    if seek_position:
-        assert streamdetails.duration, "Duration required for seek requests"
-    http_session = mass.http_session if verify_ssl else mass.http_session_no_ssl
-    # try to get filesize with a head request
-    seek_supported = streamdetails.can_seek
-    if seek_position or not streamdetails.size:
-        async with http_session.head(url, allow_redirects=True, headers=HTTP_HEADERS) as resp:
-            resp.raise_for_status()
-            if size := resp.headers.get("Content-Length"):
-                streamdetails.size = int(size)
-            seek_supported = resp.headers.get("Accept-Ranges") == "bytes"
-    # headers
-    headers = {**HTTP_HEADERS}
-    timeout = ClientTimeout(total=None, connect=30, sock_read=5 * 60)
-    skip_bytes = 0
-    if seek_position and streamdetails.size:
-        assert streamdetails.duration is not None  # for type checking
-        skip_bytes = int(streamdetails.size / streamdetails.duration * seek_position)
-        headers["Range"] = f"bytes={skip_bytes}-{streamdetails.size}"
-
-    # seeking an unknown or container format is not supported due to the (moov) headers
-    if seek_position and (
-        not seek_supported
-        or streamdetails.audio_format.content_type
-        in (
-            ContentType.UNKNOWN,
-            ContentType.M4A,
-            ContentType.M4B,
-        )
-    ):
-        LOGGER.warning(
-            "Seeking in %s (%s) not possible.",
-            streamdetails.uri,
-            streamdetails.audio_format.output_format_str,
-        )
-        seek_position = 0
-        streamdetails.seek_position = 0
-
-    # start the streaming from http
-    bytes_received = 0
-    async with http_session.get(
-        url, allow_redirects=True, headers=headers, timeout=timeout
-    ) as resp:
-        is_partial = resp.status == 206
-        if seek_position and not is_partial:
-            raise InvalidDataError("HTTP source does not support seeking!")
-        resp.raise_for_status()
-        async for chunk in resp.content.iter_any():
-            bytes_received += len(chunk)
-            yield chunk
-
-    # store size on streamdetails for later use
-    if not streamdetails.size:
-        streamdetails.size = bytes_received
-    LOGGER.debug(
-        "Finished HTTP stream for %s (transferred %s/%s bytes)",
-        streamdetails.uri,
-        bytes_received,
-        streamdetails.size,
-    )
-
-
-async def get_file_stream(
-    mass: MusicAssistant,  # noqa: ARG001
-    filename: str,
-    streamdetails: StreamDetails,
-    seek_position: int = 0,
-) -> AsyncGenerator[bytes, None]:
-    """Get audio stream from local accessible file."""
-    if seek_position:
-        assert streamdetails.duration, "Duration required for seek requests"
-    if not streamdetails.size:
-        stat = await asyncio.to_thread(os.stat, filename)
-        streamdetails.size = stat.st_size
-
-    # seeking an unknown or container format is not supported due to the (moov) headers
-    if seek_position and (
-        streamdetails.audio_format.content_type
-        in (
-            ContentType.UNKNOWN,
-            ContentType.M4A,
-            ContentType.M4B,
-            ContentType.MP4,
-        )
-    ):
-        LOGGER.warning(
-            "Seeking in %s (%s) not possible.",
-            streamdetails.uri,
-            streamdetails.audio_format.output_format_str,
-        )
-        seek_position = 0
-        streamdetails.seek_position = 0
-
-    chunk_size = get_chunksize(streamdetails.audio_format)
-    async with aiofiles.open(streamdetails.data, "rb") as _file:
-        if seek_position:
-            assert streamdetails.duration is not None  # for type checking
-            seek_pos = int((streamdetails.size / streamdetails.duration) * seek_position)
-            await _file.seek(seek_pos)
-        # yield chunks of data from file
-        while True:
-            data = await _file.read(chunk_size)
-            if not data:
-                break
-            yield data
-
-
 def _get_parts_from_position(
-    parts: list[MultiPartPath], seek_position: int
+    parts: list[MultiPartPath],
+    seek_position: int,
 ) -> tuple[list[MultiPartPath], int]:
     """Get the remaining parts list from a timestamp.
 
@@ -1320,7 +265,7 @@ def _get_parts_from_position(
         # the previous track. If we're within 2 second of the end, skip the current track
         if position + 2 >= part.duration:
             LOGGER.debug(
-                f"Skipping to the next part due to seek position being at the end: {position}"
+                f"Skipping to the next part due to seek position being at the end: {position}",
             )
             if i + 1 < len(parts):
                 return parts[i + 1 :], 0
@@ -1329,90 +274,6 @@ def _get_parts_from_position(
         return parts[i:], int(position)
 
     raise IndexError(f"Could not find any candidate part for position {seek_position}")
-
-
-async def get_multi_file_stream(
-    mass: MusicAssistant,  # noqa: ARG001
-    streamdetails: StreamDetails,
-    seek_position: int = 0,
-) -> AsyncGenerator[bytes, None]:
-    """Return audio stream for a concatenation of multiple files.
-
-    Arguments:
-    seek_position: The position to seek to in seconds
-    """
-    if not isinstance(streamdetails.path, list):
-        raise InvalidDataError("Multi-file streamdetails requires a list of MultiPartPath")
-    parts, seek_position = _get_parts_from_position(streamdetails.path, seek_position)
-    files_list = [part.path for part in parts]
-
-    # concat input files
-    temp_file = f"/tmp/{shortuuid.random(20)}.txt"  # noqa: S108
-    async with aiofiles.open(temp_file, "w") as f:
-        for path in files_list:
-            await f.write(f"file '{path}'\n")
-
-    try:
-        async for chunk in get_ffmpeg_stream(
-            audio_input=temp_file,
-            input_format=streamdetails.audio_format,
-            output_format=AudioFormat(
-                content_type=ContentType.NUT,
-                sample_rate=streamdetails.audio_format.sample_rate,
-                bit_depth=streamdetails.audio_format.bit_depth,
-                channels=streamdetails.audio_format.channels,
-            ),
-            extra_input_args=[
-                "-safe",
-                "0",
-                "-f",
-                "concat",
-                "-i",
-                temp_file,
-                "-ss",
-                str(seek_position),
-            ],
-        ):
-            yield chunk
-    finally:
-        await remove_file(temp_file)
-
-
-async def get_preview_stream(
-    mass: MusicAssistant,
-    provider_instance_id_or_domain: str,
-    item_id: str,
-    media_type: MediaType = MediaType.TRACK,
-) -> AsyncGenerator[bytes, None]:
-    """Create a 30 seconds preview audioclip for the given streamdetails."""
-    if not (music_prov := mass.get_provider(provider_instance_id_or_domain)):
-        raise ProviderUnavailableError
-    if TYPE_CHECKING:  # avoid circular import
-        assert isinstance(music_prov, MusicProvider)
-
-    # Validate that item_id corresponds to a valid item in the provider for security
-    if not await music_prov.get_item(media_type, item_id):
-        msg = f"Item {item_id} not found in provider {provider_instance_id_or_domain}"
-        raise MediaNotFoundError(msg)
-
-    streamdetails = await music_prov.get_stream_details(item_id, media_type)
-    pcm_format = AudioFormat(
-        content_type=ContentType.from_bit_depth(streamdetails.audio_format.bit_depth),
-        sample_rate=streamdetails.audio_format.sample_rate,
-        bit_depth=streamdetails.audio_format.bit_depth,
-        channels=streamdetails.audio_format.channels,
-    )
-    async for chunk in get_ffmpeg_stream(
-        audio_input=get_media_stream(
-            mass=mass,
-            streamdetails=streamdetails,
-            pcm_format=pcm_format,
-        ),
-        input_format=pcm_format,
-        output_format=AudioFormat(content_type=ContentType.AAC),
-        extra_input_args=["-t", "30"],  # cut after 30 seconds
-    ):
-        yield chunk
 
 
 async def get_silence(
@@ -1476,7 +337,9 @@ async def resample_pcm_audio(
     LOGGER.log(VERBOSE_LOG_LEVEL, f"Resampling audio from {input_format} to {output_format}")
     try:
         ffmpeg_args = get_ffmpeg_args(
-            input_format=input_format, output_format=output_format, filter_params=[]
+            input_format=input_format,
+            output_format=output_format,
+            filter_params=[],
         )
         _, stdout, stderr = await communicate(ffmpeg_args, input_audio)
         if not stdout:
@@ -1553,114 +416,6 @@ def is_grouping_preventing_dsp(player: Player) -> bool:
     return is_multiple_devices and not multi_device_dsp_supported
 
 
-def is_output_limiter_enabled(mass: MusicAssistant, player: Player) -> bool:
-    """Check if the player has the output limiter enabled.
-
-    Unlike DSP, the limiter is still configurable when synchronized without MULTI_DEVICE_DSP.
-    So in grouped scenarios without MULTI_DEVICE_DSP, the permanent sync group or the leader gets
-    decides if the limiter should be turned on or not.
-    """
-    deciding_player_id = player.player_id
-    if player.state.active_group:
-        # Syncgroup, get from the group player
-        deciding_player_id = player.state.active_group
-    elif player.state.synced_to:
-        # Not in sync group, but synced, get from the leader
-        deciding_player_id = player.state.synced_to
-    output_limiter_enabled = mass.config.get_raw_player_config_value(
-        deciding_player_id,
-        CONF_ENTRY_OUTPUT_LIMITER.key,
-        CONF_ENTRY_OUTPUT_LIMITER.default_value,
-    )
-    return bool(output_limiter_enabled)
-
-
-def get_player_filter_params(
-    mass: MusicAssistant,
-    player_id: str,
-    input_format: AudioFormat,
-    output_format: AudioFormat,
-) -> list[str]:
-    """Get player specific filter parameters for ffmpeg (if any)."""
-    filter_params = []
-
-    player = mass.players.get_player(player_id)
-    # In case this is a protocol player, their DSP config is stored
-    # under the parent (native or universal) player that wraps them.
-    dsp_player_id = player_id
-    if player and player.protocol_parent_id:
-        dsp_player_id = player.protocol_parent_id
-    dsp = mass.config.get_player_dsp_config(dsp_player_id)
-    limiter_enabled = True
-
-    if player:
-        if is_grouping_preventing_dsp(player):
-            # We can not correctly apply DSP to a grouped player without multi-device DSP support,
-            # so we disable it.
-            dsp.enabled = False
-        elif player.provider.domain == "player_group" and (
-            PlayerFeature.MULTI_DEVICE_DSP not in player.state.supported_features
-        ):
-            # This is a special case! We have a player group where:
-            # - The group leader does not support MULTI_DEVICE_DSP
-            # - But only contains a single player (since nothing is preventing DSP)
-            # We can still apply the DSP of that single player.
-            if player.state.group_members:
-                child_player = mass.players.get_player(player.state.group_members[0])
-                assert child_player is not None  # for type checking
-                dsp = mass.config.get_player_dsp_config(child_player.player_id)
-            else:
-                # This should normally never happen, but if it does, we disable DSP.
-                dsp.enabled = False
-
-        # We here implicitly know what output format is used for the player
-        # in the audio processing steps. We save this information to
-        # later be able to show this to the user in the UI.
-        player.extra_data["output_format"] = output_format
-        if player.protocol_parent_id:
-            # For protocol players, we also save the output format on the parent player,
-            # since that's where the DSP is actually applied.
-            parent_player = mass.players.get_player(player.protocol_parent_id)
-            if parent_player:
-                parent_player.extra_data["output_format"] = output_format
-
-        limiter_enabled = is_output_limiter_enabled(mass, player)
-
-    if dsp.enabled:
-        # Apply input gain
-        if dsp.input_gain != 0:
-            filter_params.append(f"volume={dsp.input_gain}dB")
-
-        # Process each DSP filter sequentially
-        for f in dsp.filters:
-            if not f.enabled:
-                continue
-
-            # Apply filter
-            filter_params.extend(filter_to_ffmpeg_params(f, input_format))
-
-        # Apply output gain
-        if dsp.output_gain != 0:
-            filter_params.append(f"volume={dsp.output_gain}dB")
-
-    conf_channels = mass.config.get_raw_player_config_value(
-        player_id, CONF_OUTPUT_CHANNELS, "stereo"
-    )
-
-    # handle output mixing only left or right
-    if conf_channels == "left":
-        filter_params.append("pan=mono|c0=FL")
-    elif conf_channels == "right":
-        filter_params.append("pan=mono|c0=FR")
-
-    # Add safety limiter at the end
-    if limiter_enabled:
-        filter_params.append("alimiter=limit=-2dB:level=false:asc=true")
-
-    LOGGER.debug("Generated ffmpeg params for player %s: %s", player_id, filter_params)
-    return filter_params
-
-
 def parse_loudnorm(raw_stderr: bytes | str) -> float | None:
     """Parse Loudness measurement from ffmpeg stderr output."""
     stderr_data = raw_stderr.decode() if isinstance(raw_stderr, bytes) else raw_stderr
@@ -1676,100 +431,10 @@ def parse_loudnorm(raw_stderr: bytes | str) -> float | None:
     return None
 
 
-async def analyze_loudness(
-    mass: MusicAssistant,
-    streamdetails: StreamDetails,
-) -> None:
-    """Analyze media item's audio, to calculate EBU R128 loudness."""
-    if await mass.music.get_loudness(
-        streamdetails.item_id,
-        streamdetails.provider,
-        media_type=streamdetails.media_type,
-    ):
-        # only when needed we do the analyze job
-        return
-
-    logger = LOGGER.getChild("analyze_loudness")
-    logger.debug("Start analyzing audio for %s", streamdetails.uri)
-
-    extra_input_args = [
-        # limit to 10 minutes to reading too much in memory
-        "-t",
-        "600",
-    ]
-    # work out audio source for these streamdetails
-    stream_type = streamdetails.stream_type
-    audio_source: str | AsyncGenerator[bytes, None]
-    if stream_type == StreamType.CUSTOM:
-        music_prov = mass.get_provider(streamdetails.provider)
-        if TYPE_CHECKING:  # avoid circular import
-            assert isinstance(music_prov, MusicProvider)
-        audio_source = music_prov.get_audio_stream(streamdetails)
-    elif stream_type == StreamType.ICY:
-        assert isinstance(streamdetails.path, str)  # for type checking
-        audio_source = get_icy_radio_stream(mass, streamdetails.path, streamdetails)
-    elif stream_type == StreamType.IN_BAND:
-        assert isinstance(streamdetails.path, str)  # for type checking
-        # Use chained OGG handler for seamless playback across logical bitstreams
-        audio_source = get_chained_ogg_stream(mass, streamdetails.path, metadata_callback=None)
-    elif stream_type == StreamType.HLS:
-        assert isinstance(streamdetails.path, str)  # for type checking
-        substream = await get_hls_substream(mass, streamdetails.path)
-        audio_source = substream.path
-    else:
-        # all other stream types (HTTP, FILE, etc)
-        if stream_type == StreamType.ENCRYPTED_HTTP:
-            assert streamdetails.decryption_key is not None  # for type checking
-            extra_input_args += ["-decryption_key", streamdetails.decryption_key]
-        if isinstance(streamdetails.path, list):
-            # multi part stream - just use a single file for the measurement
-            audio_source = streamdetails.path[1].path
-        else:
-            # regular single file/url stream
-            assert isinstance(streamdetails.path, str)  # for type checking
-            audio_source = streamdetails.path
-
-    # calculate BS.1770 R128 integrated loudness with ffmpeg
-    async with FFMpeg(
-        audio_input=audio_source,
-        input_format=streamdetails.audio_format,
-        output_format=streamdetails.audio_format,
-        audio_output="NULL",
-        filter_params=["ebur128=framelog=verbose"],
-        extra_input_args=extra_input_args,
-        collect_log_history=True,
-        loglevel="info",
-    ) as ffmpeg_proc:
-        await ffmpeg_proc.wait()
-        log_lines = ffmpeg_proc.log_history
-        log_lines_str = "\n".join(log_lines)
-        try:
-            loudness_str = (
-                log_lines_str.split("Integrated loudness")[1].split("I:")[1].split("LUFS")[0]
-            )
-            loudness = float(loudness_str.strip())
-        except (IndexError, ValueError, AttributeError):
-            LOGGER.warning(
-                "Could not determine integrated loudness of %s - %s",
-                streamdetails.uri,
-                log_lines_str or "received empty value",
-            )
-        else:
-            await mass.music.set_loudness(
-                streamdetails.item_id,
-                streamdetails.provider,
-                loudness,
-                media_type=streamdetails.media_type,
-            )
-            logger.debug(
-                "Integrated loudness of %s is: %s",
-                streamdetails.uri,
-                loudness,
-            )
-
-
 def _get_normalization_mode(
-    core_config: CoreConfig, player_config: PlayerConfig, streamdetails: StreamDetails
+    core_config: CoreConfig,
+    player_config: PlayerConfig,
+    streamdetails: StreamDetails,
 ) -> VolumeNormalizationMode:
     if not player_config.get_value(CONF_VOLUME_NORMALIZATION):
         # disabled for this player
@@ -1784,8 +449,8 @@ def _get_normalization_mode(
                 CONF_VOLUME_NORMALIZATION_RADIO
                 if streamdetails.media_type == MediaType.RADIO
                 else CONF_VOLUME_NORMALIZATION_TRACKS,
-            )
-        )
+            ),
+        ),
     )
 
     # handle no measurement available but fallback to dynamic mode is allowed
