+++ music_assistant/providers/squeezelite/provider.py
@@ -12,7 +12,7 @@ from aioslimproto.server import SlimServer
 from music_assistant_models.errors import SetupFailedError
 
 from music_assistant.constants import CONF_PORT, CONF_SYNC_ADJUST, VERBOSE_LOG_LEVEL
-from music_assistant.helpers.audio import get_mime_type, get_player_filter_params
+from music_assistant.helpers.audio import get_mime_type
 from music_assistant.helpers.util import is_port_in_use
 from music_assistant.models.player_provider import PlayerProvider
 
@@ -198,7 +198,7 @@ class SqueezelitePlayerProvider(PlayerProvider):
             child_player.display_name,
         )
 
-        output_format = await self.mass.streams.get_output_format(
+        output_format = await self.mass.streams.audio.get_output_format(
             output_format_str=fmt,
             player=child_player,
             content_sample_rate=stream.audio_format.sample_rate,  # Flow PCM sample rate
@@ -207,8 +207,8 @@ class SqueezelitePlayerProvider(PlayerProvider):
 
         async for chunk in stream.get_stream(
             output_format=output_format,
-            filter_params=get_player_filter_params(
-                self.mass, child_player_id, stream.audio_format, output_format
+            filter_params=self.mass.streams.audio.get_player_filter_params(
+                child_player_id, stream.audio_format, output_format
             )
             if child_player_id
             else None,
